### 2018-09-12 #938 @arfon
Here's the paper I compiled locally [10.21105.joss.00938.pdf](https://github.com/openjournals/joss-reviews/files/2374711/10.21105.joss.00938.pdf)
