### 2018-09-12 a6435dc
Removed zipped test data files. Remove unzip script also
Former-commit-id: 4b494da6e3d3d82d1c9bef1a4ec0ed521cd09f5a [formerly fd99101617e79d6c886ea3a321eb0c9b59f4ce3b] [formerly bdd950cce6f5697f7de76f7517397e595b3c55ec [formerly 61e3ca8cf8005eb40067ea229a9fe267a81c2b0c [formerly 9b4f3c08305c70f7765924d3699b5386d483e971] [formerly 6f73b3b8d5a7655b23deaf25ef58e0dfb9c40baa [formerly cad2f603a0ac48e3c2795a538b86a9e9abdaa9e4]]]]
Former-commit-id: d7285aefd84e534c619f5e7a8e5bf105c17d33ba [formerly 0e37dd9d36c261a2b8552a6dfa8ed770203368dd]
Former-commit-id: 3fc565aca94dca4f3f469462bf339adce040a98a
Fo
[…]
