+++ appveyor.yml
@@ -30,7 +30,6 @@ install:
   - cmd: "move tests\\config.yaml config.yaml"
 
   # Generate data
-  - cmd: "%PYTHON%\\python.exe eniric_scripts\\unzip_testdata.py"
   - cmd: "%PYTHON%\\python.exe eniric\\obsolete\\make_test_data.py"
 
   # Prepare atmosphere models
