+++ .travis.yml
@@ -31,7 +31,6 @@ before_script:
   - mv tests/config.yaml config.yaml
 
   # Generate data
-  - unzip_testdata.py
   - make_test_data.py
 
   # Prepare atmosphere models
