+++ .shippable.yml
@@ -30,7 +30,6 @@ build:
     - export MPLBACKEND="AGG"  # Non-interactive
     - mv tests/config.yaml config.yaml
     # Generate data
-    - unzip_testdata.py
     - make_test_data.py
     # Prepare atmosphere models
     - make atmos
