### 2018-09-12 bb08bf0
Removed zipped test data files. Remove unzip script also
