### 2018-09-12 50658c5
Move data from .txt to .dat files.
Former-commit-id: 80321e4c9009f8946c90ba31bbf6aeafb3cf544c [formerly 2d44a3e8e37d03a5f46f3c10145b5ef809413fa5] [formerly 0b490eea2b6e5364da28695f88e7e0b5c25f5d18 [formerly 814379cf059d5634a28eb39241b99dd584752c2f]]
Former-commit-id: b1cdb4f20e699c3ba7943d6231e7d5da778089b3 [formerly d25295b309eacd4b45443f57d33a8b72aa2ee799]
Former-commit-id: b0f8b18a446d5ce024f30f32fda22cf76444a9ee
