+++ eniric/obsolete/perge_results.sh
@@ -4,3 +4,5 @@
 # remember the original code took and hour and a half each
 rm ../data/results/*.txt
 rm ../data/resampled/*.txt
+rm ../data/results/*.dat
+rm ../data/resampled/*.dat
