### 2018-09-12 f644def
Move data from .txt to .dat files.
Former-commit-id: 80321e4c9009f8946c90ba31bbf6aeafb3cf544c [formerly 2d44a3e8e37d03a5f46f3c10145b5ef809413fa5] [formerly 0b490eea2b6e5364da28695f88e7e0b5c25f5d18 [formerly 94936cec2eeb7f1af26f8be9652e9fea1c470395 [formerly af39cd2c3b9cc70e97521c01dec69d9a30425328] [formerly 1a39f3f55d55a845d8064ce211327519ee08829f [formerly 3be2fdcae8d0f656aabb5405eb97051a99aa5fcb]]]]
Former-commit-id: b1cdb4f20e699c3ba7943d6231e7d5da778089b3 [formerly d25295b309eacd4b45443f57d33a8b72aa2ee799]
Former-commit-id: b0f8b18a446d5ce024f30f32fda22cf76444a9ee
Former-commit-id: cc224e
[…]
