### 2018-09-12 d85235a
Move data from .txt to .dat files.
Former-commit-id: 35b527dc181039609f3d540a87d6a2ff6588824a [formerly b790042f874ca193d1720e4a674830568a139dfa] [formerly d9466e74d83b21f70ea7e8ed4d12dcb25358993b [formerly e99b25fda841e2a416cf4bdf120d73ed7b0a2aba [formerly 27e2f74062bca3b38429988e73349203121bc582] [formerly 9c8a6be2722a42f26f32a19085ce72725aa71b26 [formerly bb3a467cf71b5c99d433fef807e90f09ca59b4c8]]]]
Former-commit-id: e1629a875a67fba1bb65b10bc991788614710d86 [formerly eb8970b168f6b6ec8ab07360243bb54a8131b3a9]
Former-commit-id: 4aa93486ae47b4f9fd63757bfe6780e1e0b4acff
Former-commit-id: ccf953
[…]
