### 2018-09-12 814379c
Move data from .txt to .dat files.
