### 2018-09-12 3547d05
Move data from .txt to .dat files.
