### 2018-09-12 a6b2388
Move data from .txt to .dat files.
Former-commit-id: 35b527dc181039609f3d540a87d6a2ff6588824a [formerly b790042f874ca193d1720e4a674830568a139dfa] [formerly d9466e74d83b21f70ea7e8ed4d12dcb25358993b [formerly 3547d05a78aa70f599e6ae63ccdad107ea6f5f5a]]
Former-commit-id: e1629a875a67fba1bb65b10bc991788614710d86 [formerly eb8970b168f6b6ec8ab07360243bb54a8131b3a9]
Former-commit-id: 4aa93486ae47b4f9fd63757bfe6780e1e0b4acff
