+++ data/old_fluxnorm_precision_results_2017.txt
@@ -1,181 +1,0 @@
-# id	prec_1	prec_2	prec_3
-M0-Z-1.0-60k	8.9442553	14.6927593	9.2518564
-M0-Z-1.0-80k	5.9634041	9.8199524	6.1743884
-M0-Z-1.0-100k	4.4818210	7.4444165	4.6448772
-M0-Z-5.0-60k	13.5614219	22.1395888	14.0130084
-M0-Z-5.0-80k	10.5794370	17.2240049	10.9336402
-M0-Z-5.0-100k	8.9545022	14.5881528	9.2557736
-M0-Z-10.0-60k	23.8571795	38.1953139	24.6175121
-M0-Z-10.0-80k	19.8614711	31.7434591	20.4968602
-M0-Z-10.0-100k	17.4007353	27.8331288	17.9588638
-M0-Y-1.0-60k	8.7040047	10.4154095	8.8039194
-M0-Y-1.0-80k	5.4329225	6.5123892	5.4945659
-M0-Y-1.0-100k	3.8435542	4.6132878	3.8868294
-M0-Y-5.0-60k	14.6116744	17.4079795	14.7830541
-M0-Y-5.0-80k	11.0542124	13.1724880	11.1840566
-M0-Y-5.0-100k	9.1082863	10.8534907	9.2150639
-M0-Y-10.0-60k	29.6336629	35.3925779	29.9886748
-M0-Y-10.0-80k	24.1547023	28.8808016	24.4447391
-M0-Y-10.0-100k	20.6859975	24.7239575	20.9337955
-M0-J-1.0-60k	14.5726188	38.4824294	15.3686356
-M0-J-1.0-80k	9.8130374	25.0698708	10.3342695
-M0-J-1.0-100k	7.3806081	18.5149997	7.7677064
-M0-J-5.0-60k	21.6353638	60.3861014	22.8824271
-M0-J-5.0-80k	16.8996768	46.5010842	17.8662275
-M0-J-5.0-100k	14.2840111	39.1701751	15.0974734
-M0-J-10.0-60k	37.3476583	118.6083634	39.6999514
-M0-J-10.0-80k	31.0731358	97.6930138	33.0133337
-M0-J-10.0-100k	27.1784563	85.0391102	28.8749211
-M0-H-1.0-60k	6.7407751	10.7293157	6.8822249
-M0-H-1.0-80k	4.7738551	7.5869182	4.8706949
-M0-H-1.0-100k	3.7914694	6.0194570	3.8672190
-M0-H-5.0-60k	9.6791080	15.3501913	9.8972630
-M0-H-5.0-80k	7.5697914	12.0035545	7.7376181
-M0-H-5.0-100k	6.4178031	10.1545191	6.5594408
-M0-H-10.0-60k	16.9771652	26.7310672	17.3793045
-M0-H-10.0-80k	14.0977176	22.1592111	14.4319923
-M0-H-10.0-100k	12.3237127	19.3417483	12.6148316
-M0-K-1.0-60k	13.4339568	59.6852994	14.3659071
-M0-K-1.0-80k	9.1640276	41.2615616	9.8000883
-M0-K-1.0-100k	7.0819473	31.9535539	7.5731955
-M0-K-5.0-60k	20.6640745	87.1912503	22.0930972
-M0-K-5.0-80k	15.9704885	68.1882368	17.0721232
-M0-K-5.0-100k	13.4594884	57.8580335	14.3890661
-M0-K-10.0-60k	38.2967052	154.5120258	40.9355660
-M0-K-10.0-80k	31.7298384	127.7942642	33.9297075
[…]
