### 2018-09-12 6df75eb
Add MANIFEST.in and set include_package_data=True
