### 2018-09-12 5a4859e
Install Starfish with pip link to github.
Former-commit-id: 17b4efad85aa2ec545a90f96720205b6f35c0936 [formerly 9e9c9477ea8eb3ff991646fdca84866bc3d9eb38] [formerly cf127954b55026d777bc934bc945ddc720bd505b [formerly 2d1b4fcba7534b8f2cbb3c121eeabe6bbd2e2b98 [formerly 953b6588b9cc90b7d3abfc1f8fb0efb81c268efc] [formerly 2edb4ce07a60315083183b02651add4069487d97 [formerly 5025d118a5e7d00db8395d9aa57400c85bca6935]]]]
Former-commit-id: 28476ab2ca9a20211ee11891558c9582edaba800 [formerly 3658cf57d66375b669e721e89ec3826c4fbac600]
Former-commit-id: 7811e4782f758566dc32deee2915990365bcf038
Former-commit-id:
[…]
