+++ appveyor.yml
@@ -12,12 +12,15 @@ install:
   - "%PYTHON%\\python.exe -m pip install wheel"
 
   # Installing Starfish
-  - "cd c:\\projects"
-  - "git clone https://github.com/jason-neal/Starfish.git"
-  - "cd c:\\projects\\Starfish"
-  - "%PYTHON%\\python.exe -m pip install -r requirements.txt"
-  - "git checkout windows"   # Especially configured to build on windows
-  - "%PYTHON%\\python.exe setup.py develop"
+  # This is likely to fail
+  - "%PYTHON%\\python.exe -m pip install https://github.com/iancze/Starfish/archive/master.zip#egg=Starfish"
+
+  #- "cd c:\\projects"
+  #- "git clone https://github.com/jason-neal/Starfish.git"
+  #- "cd c:\\projects\\Starfish"
+  #- "%PYTHON%\\python.exe -m pip install -r requirements.txt"
+  #- "git checkout windows"   # Especially configured to build on windows
+  #- "%PYTHON%\\python.exe setup.py develop"
 
   # Installing eniric now
   - cmd: "cd c:\\projects\\eniric"
