### 2018-09-12 a6c731f
Install Starfish with pip link to github.
Former-commit-id: 17b4efad85aa2ec545a90f96720205b6f35c0936 [formerly 9e9c9477ea8eb3ff991646fdca84866bc3d9eb38] [formerly cf127954b55026d777bc934bc945ddc720bd505b [formerly 0bb5834fc17c3ebe340d7b2b97792684d7b48fff]]
Former-commit-id: 28476ab2ca9a20211ee11891558c9582edaba800 [formerly 3658cf57d66375b669e721e89ec3826c4fbac600]
Former-commit-id: 7811e4782f758566dc32deee2915990365bcf038
