+++ requirements.txt
@@ -7,3 +7,5 @@ pandas==0.23.4
 pyyaml==3.13
 scipy==1.1.0
 tqdm==4.25.0
+# Starfish install- currently needs numpy and cython pre-installed
+# https://github.com/iancze/Starfish/archive/master.zip#egg=Starfish
