### 2018-09-12 0bb5834
Install Starfish with pip link to github.
