### 2018-09-12 3ef3c40
Install requiements first before starfish
Former-commit-id: 690b4d029aa659c73de9fa6a2b0e917aa425928a [formerly c6a80447b8add94481299ed234088fb036acdff0] [formerly 129eaec217159be5596837df145efd467476edcf [formerly 880e3c0abf01519738d38a74fea492a54fc27245 [formerly 1e60aac5ca0d7ea7381f79b0a4b1fbd5e7651e38] [formerly 8ca21576524b32b49b08940622cc15b0afb9d8fb [formerly 3f57f44d512c1a6bafe978830587b6200b5ef60d]]]]
Former-commit-id: ca3257ca0cea7775a0c10611503fc43d010ecfdd [formerly f2daec3bd77edd58fad052924f7914e59fa824bc]
Former-commit-id: 4316a4bde6f091ecb6fb09edb35707ac99cc2420
Former-commit-id:
[…]
