### 2018-09-12 5fa6698
pip install starfish setup requirements.
Former-commit-id: 6466b4772f0d44c130af16debb08046fd8728401 [formerly 5ea2fb442a57b6920a163fab921b80f48323c071] [formerly fbe7758b129f088b781cf12e88f9e80ee6439bf5 [formerly 7eceae47f923aa45f4bfd4d31f27bec4216e49f8 [formerly d060c54c2e9af74f5e9ee064a591b6ff50d0a967] [formerly 44292882dfdf401a6e81f5623f872492fff5b8ff [formerly 3da0faf33e8776f0414e8ded5d796b2f576232e9]]]]
Former-commit-id: ed0e48d84fca389d7fecb8e021d4559bf1ec92da [formerly 1e43808ca876b969124e6943a8a5490488fb3760]
Former-commit-id: a57354d7ca85c5e3a6094a14ed0bdce5d54bf129
Former-commit-id: 
[…]
