### 2018-09-12 3e46b88
pip install starfish setup requirements.
Former-commit-id: 6466b4772f0d44c130af16debb08046fd8728401 [formerly 5ea2fb442a57b6920a163fab921b80f48323c071] [formerly fbe7758b129f088b781cf12e88f9e80ee6439bf5 [formerly 64236461df1664bebeed22075e13e90361d65754]]
Former-commit-id: ed0e48d84fca389d7fecb8e021d4559bf1ec92da [formerly 1e43808ca876b969124e6943a8a5490488fb3760]
Former-commit-id: a57354d7ca85c5e3a6094a14ed0bdce5d54bf129
