+++ .travis.yml
@@ -15,6 +15,7 @@ install:
   - travis_retry pip install -r requirements_dev.txt
 
   # Install Starfish
+  - pip install numpy cython
   - pip install https://github.com/iancze/Starfish/archive/master.zip#egg=Starfish
 
   # Now install eniric
