+++ .shippable.yml
@@ -15,8 +15,8 @@ build:
     - travis_retry pip install -r requirements_dev.txt
 
     # Install Starfish
+    - pip install numpy cython
     - pip install https://github.com/iancze/Starfish/archive/master.zip#egg=Starfish
-
 
     # Now install eniric
     - python setup.py install
