### 2018-09-12 837630d
Revert appveyor to cloning environment
Former-commit-id: d975801a9ea51a08c56579a700b9a445f4e18b8f [formerly 6555654f0531ea12ab8ab64f544cf02002672014] [formerly 9111f103d19e4a1fbd6eba209eb948083067f64e [formerly 758ea207ffb030b91c4062d27ae51698814b812e]]
Former-commit-id: 167446fe74c867854cc6fc7e5c558bb997ba73f4 [formerly 5536365e60d7952349a75b9c62e69b8a143bd2d8]
Former-commit-id: 680c1528f81fa3a5176242fc3252c4b59c92359e
