+++ appveyor.yml
@@ -15,20 +15,19 @@ install:
   - "%PYTHON%\\python.exe -m pip install -r requirements_dev.txt"
 
   # Installing Starfish
-  # This is likely to fail
-  - "%PYTHON%\\python.exe -m pip install numpy cython"
-  - "%PYTHON%\\python.exe -m pip install https://github.com/iancze/Starfish/archive/master.zip#egg=Starfish"
+  # This does not work due to Starfish compilation on Windows.
+  # - "%PYTHON%\\python.exe -m pip install numpy cython"
+  # - "%PYTHON%\\python.exe -m pip install https://github.com/iancze/Starfish/archive/master.zip#egg=Starfish"
 
-  #- "cd c:\\projects"
-  #- "git clone https://github.com/jason-neal/Starfish.git"
-  #- "cd c:\\projects\\Starfish"
-  #- "%PYTHON%\\python.exe -m pip install -r requirements.txt"
-  #- "git checkout windows"   # Especially configured to build on windows
-  #- "%PYTHON%\\python.exe setup.py develop"
+  - "cd c:\\projects"
+  - "git clone https://github.com/jason-neal/Starfish.git"
+  - "cd c:\\projects\\Starfish"
+  - "%PYTHON%\\python.exe -m pip install -r requirements.txt"
+  - "git checkout windows"   # Especially configured to build on windows
+  - "%PYTHON%\\python.exe setup.py develop"
 
   # Installing eniric now
- # - cmd: "cd c:\\projects\\eniric"
-
+  - cmd: "cd c:\\projects\\eniric"
   - "%PYTHON%\\python.exe setup.py install"
 
   # Setup config and make data
