### 2018-09-12 575b1bc
wget starfish requirements and install first.
h5py missing from steup.py in Starfish.
