### 2018-09-12 76df854
wget starfish requirements and install first.
h5py missing from steup.py in Starfish.


Former-commit-id: 14adecd81e43510180f8a0974dd00aa2118bdcdb [formerly 8ac1575ea67c23f3ebf527cc4df99ff020de5855] [formerly e9c48df31543bad5d0dd6149e0dc8417b2cc0abd [formerly 575b1bcf20b66b8c3bfdd7dcea1fe309aefc15ee]]
Former-commit-id: 4a86ea2a7af078b6e4e2196cf79993e509410bf2 [formerly 6b2f6a06362b5f7b17dc95ab178dac883d73ee9f]
Former-commit-id: 886eca837a555264a02203f7aaeedd077460c5af
