### 2018-09-12 6149c93
Update pip in CI.
