+++ appveyor.yml
@@ -8,6 +8,7 @@ environment:
     - PYTHON: "C:\\Python37-x64"
 
 install:
+  - "%PYTHON%\\python.exe -m pip install --upgrade pip"
   # We need wheel installed to build wheels
   - "%PYTHON%\\python.exe -m pip install wheel"
 
