### 2018-09-12 110cdc7
Update pip in CI.
Former-commit-id: cf520247251028b258b351a165f7084459c27bb2 [formerly 66251d12e3ff1709342757d8c54ca50cd113e9f3] [formerly 322fcaf1bb7ad1bf8408b8079bf841fda20b588e [formerly 17bbc970c7d4f8f57a264275324523cd80133f15 [formerly 91720a1f330027ff39939271d6e0a9290455cc0a] [formerly 9f0cad8bf21a769db2382c4947d22a95c50701db [formerly cb105db62cf33f6fdd6626c6cb0c72d7adb9585a]]]]
Former-commit-id: 0cbf7e744f0565a34045d0d41281d1d821c1d96b [formerly 67e0f0eedf00ff6b51d72c3656b59c2cbc1ea91f]
Former-commit-id: 823b32730bca7737545c8e565117f73ea51f53f0
Former-commit-id: 2e14f577d03375b3d19d479
[…]
