### 2018-09-12 72c1ea9
Update pip in CI.
Former-commit-id: cf520247251028b258b351a165f7084459c27bb2 [formerly 66251d12e3ff1709342757d8c54ca50cd113e9f3] [formerly 322fcaf1bb7ad1bf8408b8079bf841fda20b588e [formerly 6149c9309413a05f4201c3aa4187f8eb115ba124]]
Former-commit-id: 0cbf7e744f0565a34045d0d41281d1d821c1d96b [formerly 67e0f0eedf00ff6b51d72c3656b59c2cbc1ea91f]
Former-commit-id: 823b32730bca7737545c8e565117f73ea51f53f0
