### 2018-09-12 0f888fd
xfail possible issues with Starfish installation.
Former-commit-id: 3bc8b63acbcd65bc886d0214beaa8d00eabe6fbf [formerly e80ebfb2f5baf2e108ee2f0862983ef77eeeb83c] [formerly 573cbb2ec6a7bf189288e64970f99c68c5988580 [formerly 71d90a8083d3f7d05735cbdb359f614b9c1cdea9]]
Former-commit-id: c0030b333d94ef8f82866e8a243e5472a631987d [formerly 7fc504a234bd55383761e44ceb94914308e3f3ae]
Former-commit-id: f9e8ac2e6d731649bfeae550cafa86310965bde0
