### 2018-09-12 23eac69
Remove unused plot command line flag
Former-commit-id: 73740a4f19ffeeecfa015779b24d37f42daeaf89 [formerly ffee767150a0c3a950595ed71680de5ce3852cff] [formerly 406e271df3b777bf8f84deb6c00e7cc0cd163fdb [formerly 6a31aad1ad18eb969bf60374340c72362be794bd]]
Former-commit-id: b5e5d5e58413f9d0cd6e89106953e3c585b6b1e7 [formerly d7a9aea49068f9ea08aedbf055be3b5f40fc4186]
Former-commit-id: 5f2fc8231b2fe0fd1507503f51144f3395043288
