### 2018-09-12 #938 @arfon
> @whedon generate pdf from branch develop

@jason-neal - did you manage to strip down the size of this repository?
