### 2018-09-12 #938 @jason-neal
Yea I think I cleaned out ~1.3GB from the repo. Its now just under 300M...
