### 2018-09-12 8480bd3
Bump version 1.0.0rc2
Former-commit-id: db0f17abd542062551b640402109a6e3165ab2b3 [formerly 24382f56bc0ffa0945f9e77854e43017772142b5]
Former-commit-id: 363ac372749f4439bad4fd969d6c38eed25dfca0
