### 2018-09-12 #938 @arfon
@jason-neal we only recently added the functionality to optionally build from a branch and it doesn't look like it's working properly. Could you try either merging your `develop` branch into `master` or putting the `paper.md` and other paper-related resources into the `master` branch please?
