### 2018-09-13 ec44c90
Update utilities.py
Remove dangerous lists

Former-commit-id: 47646b17b1804cf00c434d3b19711c52a7de12d9 [formerly c583b864e053ae7d665656b90eeb557b58b96244]
Former-commit-id: d808c54ddb041c5881f2879358b617ccaa91e135
