+++ eniric/utilities.py
@@ -285,7 +285,7 @@ def load_aces_spectrum(
     params: Union[ndarray, List[float]],
     photons: bool = True,
     air: bool = False,
-    wl_range: List[float] = [3000, 54000],
+    wl_range: Union[List[float], Tuple[float,float]]  = (3000, 54000),
 ):
     """Load a Phoenix spectrum from the phoenix library using STARFISH.
 
@@ -297,8 +297,8 @@ def load_aces_spectrum(
         Necessary conversions into photons for precisions.
     air: bool
        Convert to air wavelengths (default = False).
-    wl_range: [float, float]
-        Min/Max wavelength range to load. Default = [3000, 54000] Angstrom.
+    wl_range: (float, float)
+        Min/Max wavelength range to load. Default = (3000, 54000) Angstrom.
 
     Returns
     -------
@@ -350,7 +350,7 @@ def load_btsettl_spectrum(
     params: Union[ndarray, List[float]],
     photons: bool = True,
     air: bool = False,
-    wl_range: List[float] = [3000, 30000],
+    wl_range: Union[List[int], Tuple[int,int]] = (3000, 30000),
 ):
     """Load a BT-Settl spectrum from the CIFIST2011 library using STARFISH.
 
@@ -362,8 +362,8 @@ def load_btsettl_spectrum(
         Necessary conversions into photons for precisions.
     air: bool
        Convert to air wavelengths (default = False).
-    wl_range: [float, float]
-        Min/Max wavelength range to load. Default = [3000, 30000] Angstrom.
+    wl_range: (float, float)
+        Min/Max wavelength range to load. Default = (3000, 30000) Angstrom.
 
     Returns
     -------
