+++ eniric_scripts/__init__.py
@@ -1,6 +1,6 @@
 __all__ = [
     "bary_shift_atmmodel",
-    "phoenix_precision.py",
+    "phoenix_precision",
     "precision_four_panel",
     "split_atmmodel",
 ]
