### 2018-09-13 #938 @jason-neal
Proof looks ok to me. (the figure should probably shrunk to keep it to 3 pages but I from looking at other reviews that can wait until after review/acceptance.)
