### 2018-09-13 fb05a52
Try editable windows starfish install on appveyor
Former-commit-id: 160092b54e9598793ee81e920169d15b50922faf [formerly 0fea4831783e5429624130ec9c07a9629981662f]
Former-commit-id: 3b5bc96f13d57965707eba0f7317d346e371814f
