### 2018-09-13 787e287
Keep/Download data outside of repository. Download scripts and delete large files from repo.
