+++ appveyor.yml
@@ -35,7 +35,9 @@ install:
   # This is not the best place to put this but before_test didn't work.
   - cmd: "move tests\\config.yaml config.yaml"
 
-  # Generate data
+  # Downloading Eniric data with bash script
+  - ps: "sh download_eniric_data.sh"
+  # Generate data  # Should not do much as data already downloaded.
   - cmd: "%PYTHON%\\python.exe eniric\\obsolete\\make_test_data.py"
 
   # Prepare atmosphere models
