### 2018-09-13 4e514fe
Keep/Download data outside of repository. Download scripts and delete large files from repo.
Former-commit-id: e235818047bce0e30a2f0662af84af2ec1e66215 [formerly c67a6833524ef586916cfb19ba8a2fba18be60a5]
Former-commit-id: 728ac38e7133f5052e24e951a2edb1f75a57703c
