+++ eniric_scripts/download_eniric_data.sh
@@ -1,0 +1,54 @@
+#!/usr/bin/env bash
+
+# Inspired by Starfish
+# Invoke this script from the top directory.
+
+### DOWNLOAD Transmission spectrum
+
+
+DIRECTORY="data/atmmodel/"
+
+# Check to see if libraries/raw/ directory exists, if not, make it.
+if [ ! -d "$DIRECTORY" ]; then
+  echo $DIRECTORY does not exist, creating.
+  mkdir -p $DIRECTORY
+fi
+(cd $DIRECTORY
+    wget "https://www.dropbox.com/s/uab283lmkaptsib/Average_TAPAS_2014.dat.tar.gz"
+    untar_here.py Average_TAPAS_2014.dat.tar.gz
+)
+
+
+DIRECTORY2="data/test_data/"
+# Check to see if libraries/raw/ directory exists, if not, make it.
+if [ ! -d "$DIRECTORY2" ]; then
+  echo $DIRECTORY2 does not exist, creating.
+  mkdir -p $DIRECTORY2
+fi
+
+(cd $DIRECTORY2
+    wget "https://www.dropbox.com/s/oq2x7dsjeuxrf7t/obsolete.tar.gz"
+    untar_here.py obsolete.tar.gz
+    rm obsolete.tar.gz
+)
+
+
+
+DIRECTORY3="data/"
+
+# Check to see if libraries/raw/ directory exists, if not, make it.
+if [ ! -d "$DIRECTORY3" ]; then
+  echo $DIRECTORY2 does not exist, creating.
+  mkdir -p $DIRECTORY3
+fi
+
+(cd $DIRECTORY3
+    wget "https://www.dropbox.com/s/i4cxjrhcbx6e37x/precision.tar.gz"
+    untar_here.py precision.tar.gz
+    rm precision.tar.gz
+)
+
+
+# This should download the PHOENIX spectra
+download_test_PHOENIX_spec.sh
+
