### 2018-09-13 #938 @arfon
@jason-neal - could you take a look at the [list of potential reviewers](https://bit.ly/joss-reviewers) and identify some people who could review this submission? Also, feel free to tag anyone you might recommend here with their GitHub handle.
