### 2018-09-13 b611943
Update appveyor.yml
use sh command

Former-commit-id: 2f8deee4d3d7af1ad84d8219e37be0b0e6e5089e [formerly 9dd5288826d2d5e67fe08a0d5febf3cf9bb60d8d]
Former-commit-id: a4a6ddb6250164a4250e24d9a192defaed858135
