### 2018-09-14 fc9a3f3
Update appveyor.yml
Try bash, and Fast finish.

Former-commit-id: ca4db13f9dd4e0f7bd6093e57a03141a0ad60066 [formerly 381ca55580c5d74bf1fad1ead5c339325ca90d40]
Former-commit-id: 0c7229ef0273aabe3216f7816b66e1b2d68aee38
