### 2018-09-14 da289dc
Update appveyor.yml
Update script path

Former-commit-id: 4e014e533a36c18a0eb929ce0a7cbb18b5cc17dd [formerly 1ad1219385f60666a3156367a8cd27acb19143c7]
Former-commit-id: e97f63adc587d8abeb57ba2da5f115aedb6556e8
