### 2018-09-14 #938 @jason-neal
From the list possibly @nespinoza.

@mzechmeister

I also use/cite pieces from of @sczesla, @iancze but I don't know weather that causes any conflicts.
