### 2018-09-14 63fc10c
Try powershell script for appveyor/windows
