+++ appveyor.yml
@@ -35,8 +35,8 @@ install:
   # This is not the best place to put this but before_test didn't work.
   - cmd: "move tests\\config.yaml config.yaml"
 
-  # Downloading Eniric data with bash script
-  - bash "eniric_scripts\\download_eniric_data.sh"
+  # Downloading Eniric data with powershell script for windows
+  - ps: ".\eniric_scripts\\ps_download_eniric_data.ps1"
   # Generate data  # Should not do much as data already downloaded.
   - cmd: "%PYTHON%\\python.exe eniric\\obsolete\\make_test_data.py"
 
