### 2018-09-14 d3a1c80
Update appveyor.yml
missing slash
