### 2018-09-14 160e893
blacken files.
Former-commit-id: 05b81992e7dbd27f3d10b70579e23ca992217a13 [formerly 296070e62f26cbdcd63c41403c2a42fd4aab8ac3]
Former-commit-id: 8a7dea81945a2dbe65bd6f4127d8866b0ab08671
