### 2018-09-14 98e6540
Fix script path to untar_here
(so install not required for data download)


Former-commit-id: b0a2c24ab6254aeb6631adfdc56f2258101b8c19 [formerly e24da31374d3433fdd7c2ea3b4f9c1ceaed4685d]
Former-commit-id: 4afb74a3426214f300e3ee54add494689af64b50
