### 2018-09-14 376d36a
num_procs=1 skips parallel processing aswell as 0.
Former-commit-id: 40c6f34b92a43aee84b483d104010ae7ccd049f7 [formerly 021b3453ae25ce6f27d429b775cd47723760fe2a]
Former-commit-id: 3c843c0d6f981bba510d18915c76807eb04e8dc3
