### 2018-09-14 53b8f90
Move starfish pip install outside of eniric.
