+++ MANIFEST.in
@@ -2,8 +2,8 @@
 include *.md
 include LICENCE
 include requirements*.txt
-recursive-include data/ *
-
+#recursive-include data/ *
+inlcude pyproject.toml
 include tests/test_*.py
 include config.yaml
 include eniric/config.yaml
@@ -11,3 +11,5 @@ include tests/config.yaml
 include docs/Notebooks/*.ipynb
 include docs/Notebooks/config.yaml
 include eniric_scripts *
+include Makefile
+include paper/*
