### 2018-09-17 171ad2f
Update paper.md
Update acknowledgements.
