### 2018-09-21 1cfe5fe
Update download_eniric_data.sh
Avoid redirects from dropbox by requesting correct "raw" location

Former-commit-id: aa1cdb679f52b308ecdb2071de9fb81018ae355e [formerly b0ef92b4ef8c0a7aebafeb51f7e83a17f61c5514]
Former-commit-id: 4a382339776c7783889245a75c2fc1ad546aee5e
