### 2018-09-21 a4654cf
Update config.yaml
Fix default paths in backup config file.
