### 2018-09-21 8e5083b
Remove spaces in output, base precomputed check on X number of columns rather than  X number of characters.
Former-commit-id: 8dbfac052ae6aacf8c24f978c334c68bd5e90649 [formerly d8fccf66659a4deb98ec0a6538a1061971e78bab]
Former-commit-id: 32c779519d87938646339ccdcbc1b8d3d404a238
