### 2018-09-21 0a02b09
Check for artigau correction in csv also.
Former-commit-id: b612046c87e0d18de5cbc6f705a99ab46bab9c0b [formerly f31c98a97ad2597c3c278ae6a8c3fd773e1a54b9]
Former-commit-id: 601ca8cdab0e73d1519676d49d09be27a427a2be
