### 2018-09-21 0945273
Add verbose to phoenix_precision.py. (to make silent by default)
Former-commit-id: 7a4499ae795c522df3dd2b18d8a64867c44e0951 [formerly 78db56ecf82b1a4a9078228ddf5bda26e5a3e317]
Former-commit-id: 1d74583205956ce4c9b1130b975713a45236c14b
