+++ paper/paper.bib
@@ -18,6 +18,26 @@
   note = {00280}
 }
 
+@article{czekala_constructing_2015,
+  archivePrefix = {arXiv},
+  eprinttype = {arxiv},
+  eprint = {1412.5177},
+  title = {Constructing {{A Flexible Likelihood Function For Spectroscopic Inference}}},
+  volume = {812},
+  issn = {1538-4357},
+  doi = {10.1088/0004-637X/812/2/128},
+  abstract = {We present a modular, extensible likelihood framework for spectroscopic inference based on synthetic model spectra. The subtraction of an imperfect model from a continuously sampled spectrum introduces covariance between adjacent datapoints (pixels) into the residual spectrum. For the high signal-to-noise data with large spectral range that is commonly employed in stellar astrophysics, that covariant structure can lead to dramatically underestimated parameter uncertainties (and, in some cases, biases). We construct a likelihood function that accounts for the structure of the covariance matrix, utilizing the machinery of Gaussian process kernels. This framework specifically address the common problem of mismatches in model spectral line strengths (with respect to data) due to intrinsic model imperfections (e.g., in the atomic/molecular databases or opacity prescriptions) by developing a novel local covariance kernel formalism that identifies and self-consistently downweights pathological spectral line "outliers." By fitting many spectra in a hierarchical manner, these local kernels provide a mechanism to learn about and build data-driven corrections to synthetic spectral libraries. An open-source software implementation of this approach is available at http://iancze.github.io/Starfish, including a sophisticated probabilistic scheme for spectral interpolation when using model libraries that are sparsely sampled in the stellar parameters. We demonstrate some salient features of the framework by fitting the high resolution \$V\$-band spectrum of WASP-14, an F5 dwarf with a transiting exoplanet, and the moderate resolution \$K\$-band spectrum of Gliese 51, an M5 field dwarf.},
+  number = {2},
+  journal = {The Astrophysical Journal},
+  author = {Czekala, Ian and Andrews, Sean M. and Mandel, Kaisey S. and Hogg, David W. and Green, Gregory M.},
+  month = oct,
+  year = {2015},
+  keywords = {Astrophysics - Earth and Planetary Astrophysics,Astrophysics - Instrumentation and Methods for Astrophysics,Astrophysics - Solar and Stellar Astrophysics},
+  pages = {128},
[…]
