### 2018-09-21 dddd6b4
Silence other scripts.
