### 2018-09-21 ee4f3c5
Silence other scripts.
Former-commit-id: 29aa25730ce12819f3c1d5c0048d3d78ea3d0d80 [formerly 09aef4d8615ede33c5f3cebbff9fec5be8852218]
Former-commit-id: edd46e0abe517059cb4cf79e8876ad568f2b5dd8
