### 2018-09-21 517b1af
fix slpit_atmmodel
Former-commit-id: 7ac190da9c9979d44896183f059c75f3d60f2659 [formerly d8ad0976513e7a491b57d4d3b5f906530c2ee4ca]
Former-commit-id: 114bcafb04c1b68646534ef1056b10b987c249a1
