### 2018-09-21 fb5a753
Remove pytest verbose
