+++ setup.cfg
@@ -6,10 +6,10 @@
 test=pytest
 
 [tool:pytest]
-addopts = --verbose
+# addopts = --verbose
 
 [flake8]
-ignore = E501, E203,W503
+ignore = E501, E203, W503
 max-line-length = 88
 max-complexity = 20
 
