### 2018-09-21 e5ea361
Manually Downsample BT-Settl spectrum.
Former-commit-id: 9da784349ebb881a7814127b65491228673ae2ee [formerly 56d672713340ab8fff6bb499b5b6b14c43a9d468]
Former-commit-id: 230a54f231338536824f901d39b6b8a33342a288
