### 2018-09-21 c4ce192
fix equation
Former-commit-id: 188229cb4b6b6ddc3c754ebbbc184d25deca7816 [formerly 73b0f029ed2265b0422b1c6225c7563433f97734]
Former-commit-id: 1cbe898dcd0114c2d5895ce31f5e0459fa71992f
