### 2018-09-21 19f63b0
Simplify table names.
Former-commit-id: 9eead8e82f618f74677a381b617015d12a1fc094 [formerly 016a41a9656105276b5f6b4b96b4398caafd4daf]
Former-commit-id: 1934aedc0b42b6c88cee77e8b5d0ab9161aacbff
