### 2018-09-24 f2ee26c
Fix default config.yaml.
Don't pass empty list to os.path.join for phoenix and bt-settl paths
