### 2018-09-24 3f1d55c
Add more api documentation for do_analysis.
