### 2018-09-24 156bdbd
Use multiprocessing in context manager
