### 2018-09-24 5e00c12
Use a single processor Pool in phoenix_precision.py.
