### 2018-09-24 d3a550f
Don't check isinstance mprocess Pool, check int instead. (no issues that way).
Add tests for convolution, namely the num_procs type.
