### 2018-09-24 ecb62b7
Put verbose in conv_kwargs
