### 2018-09-24 1288d1b
Add maxtaskperchild to global worker pool.
