### 2018-09-24 0b1e467
Fix to keep num_procs = None also valid.
Test for num_procs=None
