+++ eniric/broaden.py
@@ -115,11 +115,11 @@ def rotational_convolution(
 
     tqdm_wav = tqdm(wavelength, disable=not verbose)
 
+    if num_procs is None:
+        num_procs = num_procs_minus_1
+
     if isinstance(num_procs, int):
         if (num_procs != 0) or (num_procs != 1):
-            if num_procs is None:
-                num_procs = num_procs_minus_1
-
             with mprocess.Pool(processes=num_procs) as mproc_pool:
                 convolved_flux = np.array(
                     mproc_pool.map(element_rot_convolution, tqdm_wav)
@@ -226,11 +226,11 @@ def resolution_convolution(
 
     tqdm_wav = tqdm(wavelength, disable=not verbose)
 
+    if num_procs is None:
+        num_procs = num_procs_minus_1
+
     if isinstance(num_procs, int):
         if (num_procs != 0) or (num_procs != 1):
-            if num_procs is None:
-                num_procs = num_procs_minus_1
-
             with mprocess.Pool(processes=num_procs) as mproc_pool:
                 convolved_flux = np.array(
                     mproc_pool.map(element_res_convolution, tqdm_wav)
