### 2018-09-24 085afef
Allow multiprocess pool passed to convolution (#100)
* Add more api documentation for do_analysis.

* import multiprocess.pool

* Use multiprocessing in context manager

* Use a single processor Pool in phoenix_precision.py.

* Don't check isinstance mprocess Pool, check int instead. (no issues that way).

* Add tests for convolution, namely the num_procs type and num_procs=None

* Put verbose in conv_kwargs

* Add maxtaskperchild to global worker pool.

* Fix outfile in ps download
