### 2018-09-27 7407dbb
Periodically flush data to file.
Former-commit-id: 2bba92603f5600408d3241042485106b0e5da2e8 [formerly 6dcd6d0ad635af519f292651426ddc26b2b01b9e]
Former-commit-id: fbe0d495e380f5d6dcec09de5a9a28afe39f77c4
