### 2018-09-27 fd8217d
Fix normalization parameter make it disable_normalization
