### 2018-09-27 3044284
Fix global multiprocess Pool setup.
Former-commit-id: e2593f046a4caac8b31aafa272a3636580c04a68 [formerly aa74ed850e56fc7449e2903500601e85b77aaeea]
Former-commit-id: fc86b7369f22c7d8d6e67c3d4bc258ca77871121
