### 2018-09-27 3096891
Fix multiprocess on 1 or 0 num_procs.
Former-commit-id: 882b164c2b6a93d57534526a849501b62052eabe [formerly d32fee99599ea60979ceccf5fd4fd775bcd487fd]
Former-commit-id: a6a371fae69b330ccf47bf91b1e6b76cba803b37
