### 2018-09-27 5115999
gloabl_pool flag.
