### 2018-09-27 7bbb285
Chunksize and max tasks.
Each worker will be replaced after handling
chunksize * maxtasksperchild pixels
