### 2018-09-27 94cd114
Open/close file for every line written.
