### 2018-09-28 223f91f
Remove global_pool from phoenix_precision.py
