### 2018-09-28 0e7c50b
Replace multiprocess with joblib as parrelization creation inside convolution.
