### 2018-09-28 741e59f
Joblib parallel valid input for convlutions.
