+++ requirements.txt
@@ -1,5 +1,5 @@
 astropy==3.0.4
-joblib==0.12.3  # >=0.12.3
+joblib==0.12.5  # >=0.12.3
 matplotlib==2.2.3
 multiprocess==0.70.6
 numpy==1.15.0
