### 2018-10-01 6807c1a
Update numpy from 1.15.0 to 1.15.2
