+++ requirements.txt
@@ -2,7 +2,7 @@ astropy==3.0.4
 joblib==0.12.5  # >=0.12.3
 matplotlib==3.0.0
 multiprocess==0.70.6
-numpy==1.15.0
+numpy==1.15.2
 pandas==0.23.4
 pyyaml==3.13
 scipy==1.1.0
