### 2018-10-01 88b7d5d
Scheduled monthly dependency update for October (#102)
* Update joblib from 0.12.3 to 0.12.5

* Update matplotlib from 2.2.3 to 3.0.0

* Update numpy from 1.15.0 to 1.15.2

* Update hypothesis from 3.69.1 to 3.74.0

* Update mypy from 0.620 to 0.630

* Update pre-commit from 1.10.5 to 1.11.1

* Update pytest from 3.7.2 to 3.8.1

* Update pytest-cov from 2.5.1 to 2.6.0
