### 2018-10-04 c5f3af2
Add python shebangs.
