+++ eniric_scripts/phoenix_precision.py
@@ -1,3 +1,4 @@
+#!/usr/bin/env python
 import argparse
 import itertools
 import os
