### 2018-10-04 e7fab9b
Use asarray
