### 2018-10-04 75fd406
Remove multiprocess. Use joblib Parallel only.
