+++ eniric/broaden.py
@@ -13,7 +13,6 @@ import joblib
 import numpy as np
 from astropy.constants import c
 from joblib import Memory, Parallel, delayed
-from multiprocess import pool
 from numpy.core.multiarray import ndarray
 from tqdm import tqdm
 
@@ -36,7 +35,7 @@ def rotational_convolution(
     *,
     epsilon: float = 0.6,
     normalize: bool = True,
-    num_procs: Optional[Union[int, pool.Pool, joblib.parallel.Parallel]] = None,
+    num_procs: Optional[Union[int, joblib.parallel.Parallel]] = None,
     verbose: bool = True,
 ) -> ndarray:
     """Perform Rotational convolution.
@@ -55,11 +54,9 @@ def rotational_convolution(
         Limb darkening coefficient
     normalize: bool (default = True)
         Area normalize the broadening kernel (corrects for unequal sampling of position).
-    num_procs: int, None or multiprocess.pool.Pool.
-        Number of processes to use with multiprocess.
+    num_procs: int, None or joblib.parallel.Parallel.
+        Number of processes to use.
         If None it is assigned to 1 less then total number of cores.
-        If num_procs = 0 or 1, then multiprocess is not used.
-        Can also be a Joblib.parallel.Parallel or multiprocess.pool.Pool instance.
     verbose: bool
         Show the tqdm progress bar (default = True).
 
@@ -119,29 +116,19 @@ def rotational_convolution(
         num_procs = num_procs_minus_1
 
     if isinstance(num_procs, int):
-        if num_procs not in [0, 1]:
-
-            with Parallel(n_jobs=num_procs) as parallel:
-                convolved_flux = np.asarray(
-                    parallel(delayed(element_rot_convolution)(wav) for wav in tqdm_wav)
-                )
-        else:  # num_procs == 0  or num_procs == 1
-            convolved_flux = np.empty_like(wavelength)  # Memory assignment
-            for ii, single_wav in enumerate(tqdm_wav):
-                convolved_flux[ii] = element_rot_convolution(single_wav)
-    elif isinstance(num_procs, joblib.parallel.Parallel):
-        convolved_flux = np.asarray(
-            num_procs(delayed(element_rot_convolution)(wav) for wav in tqdm_wav)
-        )
+        with Parallel(n_jobs=num_procs) as parallel:
+            convolved_flux = np.asarray(
+                parallel(delayed(element_rot_convolution)(wav) for wav in tqdm_wav)
+            )
     else:
         try:
-            # Assume num_procs was a multiprocess.pool.Pool
+            # Assume num_procs is a joblib.parallel.Parallel.
             convolved_flux = np.asarray(
-                num_procs.map(element_rot_convolution, tqdm_wav, chunksize=1023)
+                num_procs(delayed(element_rot_convolution)(wav) for wav in tqdm_wav)
             )
         except AttributeError:
             raise TypeError(
-                "num_proc must be an int, joblib Parallel or a multiprocess Pool. Not '{}'".format(
+                "num_proc must be an int, joblib.parallel.Parallel. Not '{}'".format(
                     type(num_procs)
                 )
             )
@@ -158,7 +145,7 @@ def resolution_convolution(
     *,
     fwhm_lim: float = 5.0,
     normalize: bool = True,
-    num_procs: Optional[Union[int, pool.Pool, joblib.parallel.Parallel]] = None,
+    num_procs: Optional[Union[int, joblib.parallel.Parallel]] = None,
     verbose: bool = True,
 ) -> ndarray:
     """Perform Resolution convolution.
@@ -177,11 +164,9 @@ def resolution_convolution(
         FWHM limit for instrument broadening.
     normalize: bool (default = True)
[…]
