### 2018-10-04 9732686
Simplify normalization / return
