### 2018-10-04 d48242b
Simplify normalization / return
Former-commit-id: 08f426379a057a9a70b2e879678501797ac59dc2 [formerly 2bcfee4bdf5861cf4d280252aa3046a2473ae0c9]
Former-commit-id: cdcbf719316acbc46b9b63e044191601ee68d528
