+++ paper/paper.bib
@@ -157,7 +157,7 @@
   volume = {612},
   issn = {0004-6361},
   doi = {10/gdmwfg},
-  abstract = {The CARMENES radial velocity (RV) survey is observing 324 M dwarfs to 
+  abstract = {The CARMENES radial velocity (RV) survey is observing 324 M dwarfs to
 search for any orbiting planets. In this paper, we present the survey
 sample by publishing one CARMENES spectrum for each M dwarf. These
 spectra cover the wavelength range 520-1710 nm at a resolution of at
@@ -197,7 +197,7 @@ level of 3-4 m s-1.},
   issn = {0004-6256},
   shorttitle = {Optical and {{Near}}-Infrared {{Radial Velocity Content}} of {{M Dwarfs}}},
   doi = {10/gddxj4},
-  abstract = {High-precision radial velocity (RV) measurements have been central in 
+  abstract = {High-precision radial velocity (RV) measurements have been central in
 the study of exoplanets during the last two decades, from the early
 discovery of hot Jupiters, to the recent mass measurements of
 Earth-sized planets uncovered by transit surveys. While optical RV is
@@ -248,8 +248,5 @@ under-predicting the H- and K-band content by half.},
   publisher = {{Cambridge University Press}},
   author = {Gray, David F.},
   year = {2005},
-  doi = {10.1017/CBO9781316036570},
-  place = {Cambridge}
+  doi = {10.1017/CBO9781316036570}
 }
-
-
