### 2018-10-08 8960ad9
Relax requirements.txt but pin all in requirements_dev.txt.
Update insecure packages only.

Import requirements.txt into setup.py.

Install requirements_dev.txt first.
