### 2018-10-08 ae8dfef
Use matrix to build py37.
