### 2018-10-08 5e5dca7
Cherry-pick update script from develop.
