+++ eniric_scripts/download_eniric_data.sh
@@ -14,7 +14,7 @@ if [ ! -d "$DIRECTORY" ]; then
   mkdir -p $DIRECTORY
 fi
 (cd $DIRECTORY
-    wget "https://www.dropbox.com/s/uab283lmkaptsib/Average_TAPAS_2014.dat.tar.gz"
+    wget "https://www.dropbox.com/s/raw/uab283lmkaptsib/Average_TAPAS_2014.dat.tar.gz"
     ../../eniric_scripts/untar_here.py Average_TAPAS_2014.dat.tar.gz
 )
 
@@ -27,7 +27,7 @@ if [ ! -d "$DIRECTORY2" ]; then
 fi
 
 (cd $DIRECTORY2
-    wget "https://www.dropbox.com/s/oq2x7dsjeuxrf7t/obsolete.tar.gz"
+    wget "https://www.dropbox.com/s/raw/oq2x7dsjeuxrf7t/obsolete.tar.gz"
     ../../eniric_scripts/untar_here.py obsolete.tar.gz
     rm obsolete.tar.gz
 )
@@ -43,12 +43,19 @@ if [ ! -d "$DIRECTORY3" ]; then
 fi
 
 (cd $DIRECTORY3
-    wget "https://www.dropbox.com/s/i4cxjrhcbx6e37x/precision.tar.gz"
+    wget "https://www.dropbox.com/s/raw/i4cxjrhcbx6e37x/precision.tar.gz"
    ../eniric_scripts/untar_here.py precision.tar.gz
     rm precision.tar.gz
 )
 
 
 # This should download the PHOENIX spectra
-download_test_PHOENIX_spec.sh
+# download_test_PHOENIX_spec.sh
 
+# FTP on travis gives several delays when getting phoenix data.
+# So putting Phoenix data in dropbox now also.
+(cd $DIRECTORY2
+    wget "https://www.dropbox.com/s/raw/skg8zwi7vnxgesj/data_raw.tar.gz"
+    ../../eniric_scripts/untar_here.py data_raw.tar.gz
+    rm data_raw.tar.gz
+)
