### 2018-10-24 f7555d3
Update bib for paper.
Former-commit-id: 54dcffd79a2c66784b1b5376605b4fa3732ad9bb [formerly e2596563d547607b25db627809c7a0707ebec7e7]
Former-commit-id: 353ba5c218f8735cdd495ac2856fa36cc158802b
