+++ paper/paper.bib
@@ -6,7 +6,7 @@
   title = {New Evolutionary Models for Pre-Main Sequence and Main Sequence Low-Mass Stars down to the Hydrogen-Burning Limit},
   volume = {577},
   issn = {0004-6361, 1432-0746},
-  doi = {10.1051/0004-6361/201425481},
+  doi = {10/gfdcsn},
   abstract = {We present new models for low-mass stars down to the hydrogen-burning limit that consistently couple atmosphere and interior structures, thereby superseding the widely used BCAH98 models. The new models include updated molecular linelists and solar abundances, as well as atmospheric convection parameters calibrated on 2D/3D radiative hydrodynamics simulations. Comparison of these models with observations in various colour-magnitude diagrams for various ages shows significant improvement over previous generations of models. The new models can solve flaws that are present in the previous ones, such as the prediction of optical colours that are too blue compared to M dwarf observations. They can also reproduce the four components of the young quadruple system LkCa 3 in a colour-magnitude diagram with one single isochrone, in contrast to any presently existing model. In this paper we also highlight the need for consistency when comparing models and observations, with the necessity of using evolutionary models and colours based on the same atmospheric structures.},
   journal = {Astronomy and Astrophysics},
   author = {Baraffe, I. and Homeier, D. and Allard, F. and Chabrier, G.},
@@ -18,18 +18,38 @@
   note = {00280}
 }
 
+@article{czekala_constructing_2015,
+  archivePrefix = {arXiv},
+  eprinttype = {arxiv},
+  eprint = {1412.5177},
+  title = {Constructing {{A Flexible Likelihood Function For Spectroscopic Inference}}},
+  volume = {812},
+  issn = {1538-4357},
+  doi = {10/gfdcsp},
[…]
