### 2018-10-29 8150b66
Update .travis.yml
Add python 3.6 to matrix
