### 2018-10-29 61475fd
Jason neal fixmatrix (#105)
* Update .travis.yml

Add python 3.6 to matrix

* Update .travis.yml
