+++ .travis.yml
@@ -1,16 +1,14 @@
 language: python
 python:
-    - 3.6
 
 matrix:
   fast_finish: true
   include:
+    - python: 3.6
+      sudo: false
     - python: 3.7
       dist: xenial  # Required for Python 3.7
       sudo: true    # travis-ci/travis-ci#9069
-    - python: 3.6
-      dist: trusty
-      sudo: false
 
 cache:
   directories:
