+++ .travis.yml
@@ -5,7 +5,7 @@ matrix:
   fast_finish: true
   include:
     - python: 3.6
-      sudo: false
+      sudo: true
     - python: 3.7
       dist: xenial  # Required for Python 3.7
       sudo: true    # travis-ci/travis-ci#9069
