+++ .travis.yml
@@ -1,11 +1,10 @@
+dist: trusty
 language: python
+sudo: false
 python:
-
+    - 3.6
 matrix:
-  fast_finish: true
   include:
-    - python: 3.6
-      sudo: true
     - python: 3.7
       dist: xenial  # Required for Python 3.7
       sudo: true    # travis-ci/travis-ci#9069
