### 2018-10-29 2d60650
Add python 3.6 again!
Former-commit-id: 292a027d5e4fe68a1132dccbef20dd54b02349e5 [formerly 5efd8c30f5c1277c5eebe95e63ea133ab65b713e]
Former-commit-id: 4a676f578aebe3a8f99c5c3405ab196541c90ea2
