+++ .travis.yml
@@ -5,6 +5,7 @@ python:
     - 3.6
 matrix:
   include:
+    - python: 3.6
     - python: 3.7
       dist: xenial  # Required for Python 3.7
       sudo: true    # travis-ci/travis-ci#9069
