+++ .travis.yml
@@ -1,14 +1,9 @@
-dist: xenial
+dist: xenial # Required for Python 3.7
+sudo: true   # travis-ci/travis-ci#9069
 language: python
-sudo: false
 python:
     - 3.6
-matrix:
-  include:
-    - python: 3.6
-    - python: 3.7
-      dist: xenial  # Required for Python 3.7
-      sudo: true    # travis-ci/travis-ci#9069
+    - 3.7
 
 cache:
   directories:
