### 2018-10-31 d2719c1
Bump and fix numpy to non-leaky version
