+++ requirements.txt
@@ -2,7 +2,7 @@ astropy==3.0.5
 joblib==0.12.5  # >=0.12.3
 matplotlib==3.0.1
 multiprocess==0.70.6.1
-numpy==1.15.3
+numpy>=1.15.4  # numpy>=1.15.4
 pandas==0.23.4
 pyyaml==3.13
 scipy==1.1.0
