### 2018-10-31 230161c
Fix end of sh file
