### 2018-10-31 #1053 @whedon
[REVIEW]: Eniric: Extended NIR Information Content
**Submitting author:** @jason-neal (<a href="http://orcid.org/0000-0003-0390-9896">Neal, Jason J.</a>)
**Repository:** <a href="https://github.com/jason-neal/eniric" target ="_blank">https://github.com/jason-neal/eniric</a>
**Version:** v1.0.0
