### 2018-10-31 #1053 @arfon
:wave: @crawfordsm - thanks again for agreeing to review this submission.

Please carry out your review in this issue by updating the checklist above and giving feedback in this issue. The reviewer guidelines are available here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
