### 2018-11-01 #1053 @jason-neal
Hi @crawfordsm.
I am currently trying to finish and submit my PhD thesis this month (November).  So I will not be readily available to address any issues immediately this month.
Just letting you know so you are aware and can plan/schedule your review accordingly.
