### 2018-11-01 a44cfc5
Unpin requirements.txt (#107)
* Relax requirements.txt but pin all in requirements_dev.txt.

Update insecure packages only.

Import requirements.txt into setup.py.

* Update Readme and Changelog
