+++ .travis.yml
@@ -12,8 +12,8 @@ cache:
 install:
   - pip install --upgrade pip
   # Install eniric requirements
-  - travis_retry pip install -r requirements.txt
   - travis_retry pip install -r requirements_dev.txt
+  - travis_retry pip install -r requirements.txt
 
   # Install Starfish
   - wget -O starfishrequirements.txt https://raw.githubusercontent.com/iancze/Starfish/master/requirements.txt
