+++ appveyor.yml
@@ -11,9 +11,9 @@ install:
   - "%PYTHON%\\python.exe -m pip install --upgrade pip"
   # We need wheel installed to build wheels
   - "%PYTHON%\\python.exe -m pip install wheel"
-
-  - "%PYTHON%\\python.exe -m pip install -r requirements.txt"
+  -
   - "%PYTHON%\\python.exe -m pip install -r requirements_dev.txt"
+  - "%PYTHON%\\python.exe -m pip install -r requirements.txt"
 
   # Installing Starfish
   - "cd c:\\projects"
