### 2018-11-01 367c241
Unpin requirements.txt (#107)
* Relax requirements.txt but pin all in requirements_dev.txt.

Update insecure packages only.

Import requirements.txt into setup.py.

* Update Readme and Changelog


Former-commit-id: 4be5b70d1d4e26510cbdc130a6282a2fcf0bab26 [formerly 7adb5f2ed656e2e85ff0436534da3ea7ed9cf030]
Former-commit-id: e8ab1570b3156a8a3bda26738413fbf8d94f3660
