### 2018-11-01 5f4e1d5
Pin matplotlib to latest version 3.0.1
