+++ requirements.txt
@@ -1,6 +1,6 @@
 astropy>=3.0.0
 joblib>=0.12.5  # >=0.12.3
-matplotlib
+matplotlib==3.0.1
 multiprocess
 numpy
 pandas
