### 2018-11-01 61f2b82
Pin numpy to latest version 1.15.3
