+++ requirements.txt
@@ -2,7 +2,7 @@ astropy>=3.0.0
 joblib>=0.12.5  # >=0.12.3
 matplotlib==3.0.1
 multiprocess==0.70.6.1
-numpy
+numpy==1.15.3
 pandas
 pyyaml
 scipy
