+++ requirements.txt
@@ -3,7 +3,7 @@ joblib>=0.12.5  # >=0.12.3
 matplotlib==3.0.1
 multiprocess==0.70.6.1
 numpy==1.15.3
-pandas
+pandas==0.23.4
 pyyaml
 scipy
 tqdm
