### 2018-11-01 5fdbf57
Pin pyyaml to latest version 3.13
