### 2018-11-01 5e37ce8
Pin scipy to latest version 1.1.0
