### 2018-11-01 cf1b4c7
Pin tqdm to latest version 4.28.1
