+++ requirements.txt
@@ -6,6 +6,6 @@ numpy==1.15.3
 pandas==0.23.4
 pyyaml==3.13
 scipy==1.1.0
-tqdm
+tqdm==4.28.1
 # Starfish install- currently needs numpy and cython pre-installed
 # https://github.com/iancze/Starfish/archive/master.zip#egg=Starfish
