### 2018-11-02 5c72bd0
Refactor script directory  ->
