### 2018-11-05 17da76a
Update requirements.txt
remove pins
