+++ requirements.txt
@@ -1,11 +1,11 @@
 astropy>=3.0.0
-joblib>=0.12.5  # >=0.12.3
-matplotlib==3.0.1
-multiprocess==0.70.6.1
-numpy==1.15.3
-pandas==0.23.4
-pyyaml==3.13
-scipy==1.1.0
-tqdm==4.28.1
+joblib>=0.12.3  # >=0.12.3
+matplotlib
+multiprocess
+numpy
+pandas
+pyyaml
+scipy
+tqdm
 # Starfish install- currently needs numpy and cython pre-installed
 # https://github.com/iancze/Starfish/archive/master.zip#egg=Starfish
