### 2018-11-05 075d397
Bump and fix numpy to non-leaky version (#106)
Former-commit-id: 80fffe470bc5864c230eecaf1471503d62b64165 [formerly d946d2a2569f8ee0f17370d7e350052db1a32f58]
Former-commit-id: ec48c3276a24074097f69731ee66bebf179c0a8d
