+++ requirements.txt
@@ -1,8 +1,8 @@
 astropy>=3.0.0
-joblib>=0.12.5  # >=0.12.3
+joblib>=0.12.3  # >=0.12.3
 matplotlib
 multiprocess
-numpy
+numpy>=1.15.4  # numpy>=1.15.4
 pandas
 pyyaml
 scipy
