### 2018-11-05 0cce554
Refactor script directory  -> (#109)
Former-commit-id: b52520bebc5c1e33a2c8d71a50ed0b7d36fb801c [formerly 91f3de80873f438c41db6c2cc929f116ce484aa1]
Former-commit-id: e9b4a3e70f9367a1757aea9758a0171c0751f0fc
