+++ eniric/obsolete/make_test_data.py
@@ -13,7 +13,7 @@ from eniric.obsolete.prepare_data import main as prepare_data
 if __name__ == "__main__":
     print("Eniric paths: {}".format(eniric.paths))
 
-    # subprocess.call("python eniric_scripts/prepare_data.py -s M0 M3 M6 M9 -l 4.50 -m 0.0", shell=True)
+    # subprocess.call("python scripts/prepare_data.py -s M0 M3 M6 M9 -l 4.50 -m 0.0", shell=True)
     prepare_data(
         startype=["M0", "M3", "M6", "M9"],
         temp=[],
@@ -33,7 +33,7 @@ if __name__ == "__main__":
     ]  # type: List[Tuple[List[str], List[str], List[int], List[str]]]
 
     for (sptype, band, vel, res) in parameters:
-        # subprocess.call(["python eniric_scripts/nIR_run.py
+        # subprocess.call(["python scripts/nIR_run.py
         #       -s {0} -b {1} -R {2} -v {3}".format(sptype, band, res, vel)], shell=True)
 
         nir_run(startype=sptype, vsini=vel, resolution=res, band=band, verbose=False)
