+++ tests/test_snr_normalization.py
@@ -3,7 +3,7 @@ import pytest
 
 import eniric.snr_normalization as snrnorm
 import eniric.utilities as utils
-from eniric_scripts.phoenix_precision import convolve_and_resample
+from scripts.phoenix_precision import convolve_and_resample
 
 xfail = pytest.mark.xfail
 
@@ -62,7 +62,9 @@ def test_band_snr_norm(testing_spectrum, sampling):
 
     assert snrnorm.snr_constant_band(
         wav, flux, band="J", snr=100, sampling=sampling, verbose=True
-    ) != snrnorm.snr_constant_wav(wav, flux, wav_ref=1.24, snr=100, sampling=sampling, verbose=True)
+    ) != snrnorm.snr_constant_wav(
+        wav, flux, wav_ref=1.24, snr=100, sampling=sampling, verbose=True
+    )
 
 
 def test_sampling_index():
@@ -136,7 +138,7 @@ def test_snr_normalization_logic(band, verbose):
     flux = 3 * np.ones(size)
     band_const = snrnorm.snr_constant_band(wav, flux, snr=3, band=band, verbose=verbose)
     wav_const = snrnorm.snr_constant_wav(
-        wav, flux, snr=3, wav_ref=(lim[0] + lim[1]) / 2, verbose=verbose,
+        wav, flux, snr=3, wav_ref=(lim[0] + lim[1]) / 2, verbose=verbose
     )
     assert band_const == 1
     assert wav_const == 1
