+++ scripts/download_eniric_data.sh
@@ -15,7 +15,7 @@ if [ ! -d "$DIRECTORY" ]; then
 fi
 (cd $DIRECTORY
     wget "https://www.dropbox.com/s/raw/uab283lmkaptsib/Average_TAPAS_2014.dat.tar.gz"
-    ../../eniric_scripts/untar_here.py Average_TAPAS_2014.dat.tar.gz
+    ../../scripts/untar_here.py Average_TAPAS_2014.dat.tar.gz
 )
 
 
@@ -28,7 +28,7 @@ fi
 
 (cd $DIRECTORY2
     wget "https://www.dropbox.com/s/raw/oq2x7dsjeuxrf7t/obsolete.tar.gz"
-    ../../eniric_scripts/untar_here.py obsolete.tar.gz
+    ../../scripts/untar_here.py obsolete.tar.gz
     rm obsolete.tar.gz
 )
 
@@ -44,7 +44,7 @@ fi
 
 (cd $DIRECTORY3
     wget "https://www.dropbox.com/s/raw/i4cxjrhcbx6e37x/precision.tar.gz"
-   ../eniric_scripts/untar_here.py precision.tar.gz
+   ../scripts/untar_here.py precision.tar.gz
     rm precision.tar.gz
 )
 
@@ -56,6 +56,6 @@ fi
 # So putting Phoenix data in dropbox now also.
 (cd $DIRECTORY2
     wget "https://www.dropbox.com/s/raw/skg8zwi7vnxgesj/data_raw.tar.gz"
-    ../../eniric_scripts/untar_here.py data_raw.tar.gz
+    ../../scripts/untar_here.py data_raw.tar.gz
     rm data_raw.tar.gz
 )
