+++ README.md
@@ -10,8 +10,8 @@
 [![PyPI version](https://badge.fury.io/py/eniric.svg)](https://badge.fury.io/py/eniric)
 
 `Eniric` is a Python 3 software to compute the theoretical Radial Velocity (RV) precision of stellar spectra.
-`Eniric` is an overhaul and extension to the code used in [Figueria et al. 2016](http://dx.doi.org/10.1051/0004-6361/201526900) to analysis the precision of M-dwarf stars. 
-Extending the performance and usability, it is able to be used on any synthetic spectra from the [PHOENIX-ACES](http://phoenix.astro.physik.uni-goettingen.de) and [BT-Settl](https://phoenix.ens-lyon.fr/Grids/BT-Settl/CIFIST2011_2015/FITS/) (CIFIST2001-2015) libraries. 
+`Eniric` is an overhaul and extension to the code used in [Figueria et al. 2016](http://dx.doi.org/10.1051/0004-6361/201526900) to analysis the precision of M-dwarf stars.
+Extending the performance and usability, it is able to be used on any synthetic spectra from the [PHOENIX-ACES](http://phoenix.astro.physik.uni-goettingen.de) and [BT-Settl](https://phoenix.ens-lyon.fr/Grids/BT-Settl/CIFIST2011_2015/FITS/) (CIFIST2001-2015) libraries.
 
 Checkout the [wiki here](https://github.com/jason-neal/eniric/wiki)!
 
@@ -59,7 +59,7 @@ see [here](https://github.com/iancze/Starfish) for more information about instal
 For **windows** installation of `Starfish` also see [starfish#87](see https://github.com/iancze/Starfish/issues/87).
 
 ##### Requirements for `Eniric` :
-These should be automatically installed (if not present) when installing `eniric`. 
+These should be automatically installed (if not present) when installing `eniric`.
 - astropy
 - joblib>=0.12.3
 - matplotlib
@@ -70,7 +70,7 @@ These should be automatically installed (if not present) when installing `eniric
 - scipy
 - tqdm
 
-If want to use known working pinned versions they can be found in `requirements_dev.txt`. 
+If want to use known working pinned versions they can be found in `requirements_dev.txt`.
 To install these run `pip install -r requirements_dev.txt`.
 
 ##### Extra requirements for Starfish:
@@ -79,25 +79,25 @@ To install these run `pip install -r requirements_dev.txt`.
 - h5py
 - scikit-learn
 
-If you are not going to use `eniric` to analyze PHOENIX-ACES or BT-Settl synthetic spectral models then you may 
+If you are not going to use `eniric` to analyze PHOENIX-ACES or BT-Settl synthetic spectral models then you may
 get away with not installing it (some tests will just xfail).
 
 
 ## Preparation
 #### Data download
-To download the data for eniric, an atmopsheric transmission spectra and some test Spectra run the 
+To download the data for eniric, an atmopsheric transmission spectra and some test Spectra run the
 following from the main repo directory
 
 Linux:
 `download_eniric_data.sh`
 
 Windows:
-`... ` 
+`... `
 
 This should place the data in `data/atmmodel` and `data/testdata` where it can be found for testing.
 
 ### Configuration
-`Eniric` uses a `config.yaml` file which is required in directory where you are running `Eniric`. (i.e. the current directory) 
+`Eniric` uses a `config.yaml` file which is required in directory where you are running `Eniric`. (i.e. the current directory)
 to specify some paths, such as the location the the synthetic spectral library.
 
 ```
@@ -110,7 +110,7 @@ The paths can either be a string or a list of strings to pass to `os.path.join` 
 
[…]
