### 2018-11-05 92698e9
Refactor script directory  -> (#109)
