+++ setup.py
@@ -43,13 +43,13 @@ config = {
         "dev": ["check-manifest"],
         "test": ["coverage", "pytest", "pytest-cov", "python-coveralls", "hypothesis"],
     },  # $ pip install -e .[dev,test]
-    "packages": ["eniric", "eniric_scripts", "eniric.obsolete"],
+    "packages": ["eniric", "scripts", "eniric.obsolete"],
     "scripts": [
-        "eniric_scripts/phoenix_precision.py",
-        "eniric_scripts/split_atmmodel.py",
-        "eniric_scripts/bary_shift_atmmodel.py",
-        "eniric_scripts/untar_here.py",
-        "eniric_scripts/download_eniric_data.sh",
+        "scripts/phoenix_precision.py",
+        "scripts/split_atmmodel.py",
+        "scripts/bary_shift_atmmodel.py",
+        "scripts/untar_here.py",
+        "scripts/download_eniric_data.sh",
         "tests/download_test_PHOENIX_spec.sh",
         "eniric/obsolete/make_test_data.py",
         "eniric/obsolete/nIR_run.py",
