### 2018-11-16 #1053 @crawfordsm
HI @jason-neal  Thanks for the information and good luck with your thesis.   I just wanted to let you know that I have started to review the code and I hope to complete it over the next couple of days.

When I do need to open an issues, do you prefer for me to do so as one single issue related to the review or multiple issues in your repository?
