### 2018-11-16 #1053 @jason-neal
Multiple issues is fine.
