### 2018-11-16 #1053 @crawfordsm
`Eniric` is software to compute the theoretical precision of stellar radial velocity computations for stellar spectra.  It is an updated python version of the code presented in Figueria et al (2016) .  The software is timely and useful and would be worthy of publication in JOSS after some revisions.   The code comes with a suite of tests, notebooks demonstrating its usage, and scripts to simplify the process of running the software. The code is able to reproduce the results in Figueria and comes with a number of additional useful features.

The major issue with the submission is that the documentation appears to be incomplete.   While the installation, configuration, and basic usage are included in the README, the information on the README is incomplete on the full usage of the package and the further documentation that is part of the wiki does not fill in this information.     Furthermore, there are a set of Jupyter notebooks detailing the usage, and while they have different levels of quality, they are not well linked to from the documentation.     Further documentation with examples would also help verify the results with the Figueria paper.

The functionality that is part of `Eniric` to calculate the theoretical signal to noise measurements would be useful beyond calculating the S/N and precision measurements.    It would be a significant benefit to provide more information to the user about how to use this functionality.    Using tools like sphinx and read the docs,  the documentation and API can be built automatically, and I would highly recommend moving from the wiki based documentation to an automatically built documentation.  This would also make the well written doc strings more accessible.  If you are unfamiliar with these, I'm happy to recommend some resource with more information.

Overall, the code and repository look appropriate for publication and provide a useful tool for calculating the theoretical limit on the precision of RV calculations.   The additional update of the documentation will make the code more accessible and useful to the community.

Minor comments will be opened in the repository and linked to this issue.
