### 2018-11-16 jason-neal/eniric#110 @crawfordsm
Edits to JOSS paper
These are a short number of edits to the JOSS paper:

Paragraph 1 With recent high-precision spectrographs targeting radial velocity (RV) precision at the 10\ cms$^{-1}$ level\ [e.g. @pepe_espresso_2014] in the quest to find the smallest planets, it is important to understand…

Paragraph 6:remove comma after ‘slices’

Paragraph 7: remove comma after ‘photon noise’

Paragraph 7 is unclear —

Paragraph 8: The first sentence is unclear

Paragraph 8: Can you spell out what the three separate conditions are?

Link: https://github.com/openjournals/joss-reviews/issues/1053
