### 2018-11-16 jason-neal/eniric#111 @crawfordsm
Citations in JOSS paper
Can you include a section in the paper with acknowledging the software used in this work and citations to those packages?

Link: https://github.com/openjournals/joss-reviews/issues/1053
