### 2018-11-16 jason-neal/eniric#112 @crawfordsm
Link to documentation
This is minor but it would be useful to add a link to the documentation from the top of the github repository.

Link: https://github.com/openjournals/joss-reviews/issues/1053
