### 2018-11-16 jason-neal/eniric#113 @crawfordsm
Starfish installation
In the installation instructions, here are some suggestions to help with the starfish installation:

1. Update the link from git@github.com:iancze/Starfish.git to the https address.
2. One possible thing to try is to include pip installing from github for starfish -- this might simplify your installation instructions.
3. If not, then you probably want to include a step to `pip install -r requirements.txt` for the starfish installation

Link: https://github.com/openjournals/joss-reviews/issues/1053
