### 2018-11-16 jason-neal/eniric#114 @crawfordsm
Installation instructions
Wiki installation and configuration:  There is slightly different information between the two.   Suggest keeping one set of docs rather than two or making sure they are both updated simultaneously.

The data download for example is only in the READMe and not the wiki though that is very useful information

Link: https://github.com/openjournals/joss-reviews/issues/1053
