### 2018-11-16 jason-neal/eniric#116 @crawfordsm
Bug in example usage
In the README example, there is an extra comma after the value for the surface gravity

Link: https://github.com/openjournals/joss-reviews/issues/1053
