### 2018-11-16 jason-neal/eniric#117 @crawfordsm
Instructions for how to run the tests
It would be useful to add instructions for how to run the tests and what the expected outputs are.

Link: https://github.com/openjournals/joss-reviews/issues/1053
