### 2018-11-16 jason-neal/eniric#120 @crawfordsm
Documentation link to example directory? 
The examples directory was very useful to take a look at things.  It would be useful to have a high level description and page linking to each of the notebooks and explaining what they do.   More consistent description of each of the notebooks would also make the notebooks more useful.

Link: https://github.com/openjournals/joss-reviews/issues/1053
