### 2018-11-16 #1053 @crawfordsm
Those are the minor comments -- thanks for writing the code and let me know if you have any questions or if anything was not clear.
