### 2018-11-17 #1053 @jason-neal
Thanks for the review.
Yes, I am aware that the documentation is incomplete and I will take the advice to move it to an automated build on Read the Docs.
I have used Sphinx and Read the Docs previously in 2016 but if you have a favourite or up-to-date resource (other than the basic sphinx/read the doc documentation) on the matter that would be helpful.

I will attempt to get this implemented over December. (Currently about to move from Portugal back to New Zealand)
