### 2018-11-19 51f777c
Edits to JOSS paper following #110
