### 2018-12-01 01100cd
Pin multiprocess to latest version 0.70.6.1
