+++ requirements.txt
@@ -1,7 +1,7 @@
 astropy>=3.0.0
 joblib>=0.12.3  # >=0.12.3
 matplotlib==3.0.2
-multiprocess
+multiprocess==0.70.6.1
 numpy>=1.15.4  # numpy>=1.15.4
 pandas
 pyyaml
