+++ requirements.txt
@@ -3,7 +3,7 @@ joblib>=0.12.3  # >=0.12.3
 matplotlib==3.0.2
 multiprocess==0.70.6.1
 numpy>=1.15.4  # numpy>=1.15.4
-pandas
+pandas==0.23.4
 pyyaml
 scipy
 tqdm
