### 2018-12-01 b195e48
Pin pyyaml to latest version 3.13
