### 2018-12-01 a16a75f
Pin scipy to latest version 1.1.0
