### 2018-12-01 60eb4bf
Pin tqdm to latest version 4.28.1
