### 2018-12-11 c49b93e
Basic sphinx quickstart configuration.
