+++ docs/source/index.rst
@@ -1,0 +1,20 @@
+.. Eniric documentation master file, created by
+   sphinx-quickstart on Tue Dec 11 23:58:47 2018.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Welcome to Eniric's documentation!
+==================================
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+
+
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
