### 2018-12-11 5e4a16f
Basic sphinx quickstart configuration.
Former-commit-id: 421facc4c25b3e895e7a8ab23206af9f1ffd59bc [formerly b0c014b0bbcd27c3a28ff56244f995d99ba4db11]
Former-commit-id: fb7cca2ad46405ce1c4ef4287e26481b703eb909
