### 2018-12-11 c70fc6c
Add .readthedocs.yml configuration.
Former-commit-id: 34777d1c0333ac2f2236078b957e8f2696f5a979 [formerly a790825746c7cae54fdb89a9c650a2f277ab5c1d]
Former-commit-id: 59ce6f4f8928b3033edbb047983f67587e2f65b6
