### 2018-12-11 fa6a2d6
Add suport and licence to docs.
Update licence to 2019 also


Former-commit-id: 07b77129497e25dfa1e6d1a82928014c974058da [formerly 199014b25b9a45b9e83818950816d757a328909a]
Former-commit-id: d219a84ec00b87c99a5fc8fe836d17870cdc4e92
