### 2018-12-11 29e239e
Add summary and features to docs
Former-commit-id: ad7649f7a155467c08a95b6978537dcf6a1371bf [formerly ad8bd63cb612959e382c8205ef1b414153cf185e]
Former-commit-id: 521db0576c370c2bc1c2eec9250b3bc04cfb074c
