### 2018-12-11 248521a
Change build image for python 3.6
Former-commit-id: 1a60bb21c89b74af25d1dee805f92057f88166be [formerly fa9d2900aed23cf1ab05a2ee124c07c6b22465ff]
Former-commit-id: ae826cd887717f8dadc0b8617336afe163493685
