+++ .readthedocs.yml
@@ -8,6 +8,9 @@ formats:
 requirements_file:
     requirements_dev.txt
 
+build:
+    image: latest
+
 python:
     version: 3.6
     setup_py_install: true
