### 2018-12-13 967f634
Use c.to("km/s") for speed of light
and re-standardize astropy constants imports.
