### 2018-12-14 560885c
Lazy load configuration using _config.py and testing.
