+++ eniric/__init__.py
@@ -8,43 +8,22 @@ __all__ = [
     "utilities",
 ]
 
-# Read the users config.yaml file.
-# If it doesn't exist, print a useful help message
 
 import os
+import warnings
 
-import yaml
+from ._config import Config
 
-try:
-    with open("config.yaml") as f:
-        config = yaml.safe_load(f)
-except FileNotFoundError as e:
-    default = __file__[:-11] + "config.yaml"
-    import warnings
-
+if os.path.exists("config.yaml"):
+    config = Config("config.yaml")
+else:
+    base_dir = os.path.dirname(__file__)
+    default = os.path.join(base_dir, "config.yaml")
     warnings.warn(
         "Using the default config.yaml file located at {0}. "
         "This is likely NOT what you want. Please create a similar "
         "'config.yaml' file in your current working directory.".format(default),
         UserWarning,
     )
-    with open(default) as f:
-        config = yaml.safe_load(f)
+    config = Config(default)
 
-# Read the YAML variables into package-level dictionaries to be used by the other programs.
-name = config["name"]
-paths = config["paths"]
-bands = config["bands"]
-custom_bands = config["custom_bands"]
-atmmodel = config["atmmodel"]
-cache = config["cache"]
-
-# Turn list into path
-for key, value in paths.items():
-    if isinstance(value, list):
-        paths[key] = os.path.join(*value)
-
-if (cache["location"] is None) or (cache["location"] == "None"):
-    cache["location"] = None  # Disables caching
-else:
-    cache["location"] = os.path.join(*cache["location"])
