### 2018-12-14 ac28193
fix version in __init__ and __all__
