### 2018-12-15 2498e3f
Remove starfish test config ad fix config tests.
