### 2018-12-15 95ef7ba
Test lazy load. xfail as not working correctly.
