### 2018-12-15 eeff318
Lazy load config. (#124)
* Use c.to("km/s") for speed of light

and re-standardize astropy constants imports.

* Lazy load configuration using _config.py and testing.

* fix version in __init__ and __all__

* Remove starfish test config ad fix config tests.

* Test lazy load. xfail as not working correctly.

* Fix cache location config.

* Lint
