+++ scripts/precision_four_panel.py
@@ -445,7 +445,7 @@ def cumulative_plot(
 if __name__ == "__main__":
     # ToDo: Add an argparse cli
 
-    precision_file = os.path.join(eniric.paths["precision"], "precision_results.csv")
+    precision_file = os.path.join(eniric.config.paths["precision"], "precision_results.csv")
     plot_precision(precision_file, teffs=[3900, 3500, 2800, 2600])
     # plot_precision(precision_file, teffs=[3900, 3500, 2800, 2600], logg=4, fe_h=1)
 
