### 2018-12-15 9329902
Lazy load config. (#124)
* Use c.to("km/s") for speed of light

and re-standardize astropy constants imports.

* Lazy load configuration using _config.py and testing.

* fix version in __init__ and __all__

* Remove starfish test config ad fix config tests.

* Test lazy load. xfail as not working correctly.

* Fix cache location config.

* Lint


Former-commit-id: f03f58005c1881691e6a557124b1d599138f02e7 [formerly 4547d2da1c4a5bcf6716089efaac911be3f2528f]
Former-commit-id: f595c84c7203da9f2b7a850c317432d095c71e94
