### 2018-12-30 8e046a1
Link and badge to read the docs documentation.
Fixes #112
