### 2018-12-30 4a745d7
Link and badge to read the docs documentation.
Fixes #112


Former-commit-id: c94b115e1a1ce0984bf880c8d36c55a9709af8fe [formerly c209922d68eac9aeaa5e170b8388c5d49c9ba12d]
Former-commit-id: 60f716cf6bc8ac24204253e88d75b4ea113f7faa
