### 2018-12-30 83a1e74
Convert documentation from wiki to rst and add to docs.
Former-commit-id: ad48891e7d2d5de2ebae3e280c807e7deddb4a06 [formerly ff0f3c4b8d9bcac839f243e0aee7a02dafbe017d]
Former-commit-id: f65a3ecf720004ab0c8a857c65a144ac3fc6d411
