+++ docs/source/index.rst
@@ -3,6 +3,31 @@
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
 
+.. image:: https://img.shields.io/badge/python-3.6+-blue.svg
+   :target: https://www.python.org/downloads/release/python-360/
+   :alt: Python 3.6+
+
+
+.. image:: https://travis-ci.org/jason-neal/eniric.svg?branch=master
+   :target: https://travis-ci.org/jason-neal/eniric
+   :alt: Build Status
+
+
+.. image:: https://coveralls.io/repos/github/jason-neal/eniric/badge.svg?branch=master
+   :target: https://coveralls.io/github/jason-neal/eniric?branch=master
+   :alt: Coverage Status
+
+
+.. image:: https://api.codacy.com/project/badge/Grade/24d3d525a79d4ae493de8c527540edef
+   :target: https://www.codacy.com/app/jason-neal/eniric?utm_source=github.com&utm_medium=referral&utm_content=jason-neal/eniric&utm_campaign=badger
+   :alt: Codacy Badge
+
+
+.. image:: https://badge.fury.io/py/eniric.svg
+   :target: https://badge.fury.io/py/eniric
+   :alt: PyPI version
+
+
 Welcome to Eniric's documentation!
 ==================================
 
@@ -10,33 +35,64 @@ Welcome to Eniric's documentation!
    :maxdepth: 2
    :caption: Contents:
 
-   installation
+   Installation
+   Configuration
+   Basic-usage
+   Broadening
+   Atmospheric-Transmission
+   Theoritical-Precision-of-PHONEIX-ACES-Spectra
 
-`Eniric` is a Python 3 software to compute the theoretical Radial Velocity (RV) precision of stellar spectra.
-`Eniric` is an overhaul and extension to the code used in `Figueria et al. 2016 <http://dx.doi.org/10.1051/0004-6361/201526900>`_ to analysis the precision of M-dwarf stars.
+
+`Eniric` is a Python 3.6+ software written to calculate the theoretical Radial Velocity (RV) precision of Near-InfraRed (NIR) stellar spectra.
+`Eniric` is an overhaul and extension to the code initially used in `Figueria et al. 2016 <http://dx.doi.org/10.1051/0004-6361/201526900>`_ to analysis the precision of M-dwarf stars in the NIR, extending its ability to use any synthetic spectra from the PHOENIX-ACES and BT-Settl libraries, or user provided spectra, and making it easier/faster to use.
 Extending the performance and usability, it is able to be used on any synthetic spectra from the `PHOENIX-ACES <http://phoenix.astro.physik.uni-goettingen.de>`_ and `BT-Settl <https://phoenix.ens-lyon.fr/Grids/BT-Settl/CIFIST2011_2015/FITS/>`_ (CIFIST2001-2015) libraries.
 
-Checkout the `wiki here <https://github.com/jason-neal/eniric/wiki>`_!
+To get started see the `installation instructions <https://github.com/jason-neal/eniric/wiki/Installation>`_\ , or the `basic usage <https://github.com/jason-neal/eniric/wiki/Basic-usage>`_.
+
 
 Features
 --------
+
 `Eniric` contains a number of features to transform and prepare the spectra (observed and synthetic).
-- `Spectral broadening <https://github.com/jason-neal/eniric/wiki/Broadening>`_
-   - Rotational
-   - Instrumental
-- `Atmospheric transmission masking <https://github.com/jason-neal/eniric/wiki/Atmospheric-Transmission>`_
-- Relative RV precision
-  - The RV precision can be calculated relative to a specified SNR per pixel in the center of a spectroscopic band.
[…]
