### 2018-12-30 63f9f9b
Convert documentation from wiki to rst and add to docs.
