### 2019-01-01 3e9c681
Move some requirements to "ci" extras,
pip install extra_requirements in CI runs.
