### 2019-01-01 25ea727
Move some requirements to "ci" extras,
pip install extra_requirements in CI runs.


Former-commit-id: d8205bccab083e35f8c2dd5b2ce7f2ea8e7d595e [formerly 48efe89cdace9d994a15a611580e6b72e1b72555]
Former-commit-id: a8c0b49111827d9047fc6dea3c55f0d2765752bf
