+++ appveyor.yml
@@ -34,7 +34,7 @@ install:
 
   # Installing eniric now
   - cmd: "cd c:\\projects\\eniric"
-  - "%PYTHON%\\python.exe setup.py install"
+  - "%PYTHON%\\python.exe -m pip install .[dev,test,docs,ci]"
 
   # Setup config and make data
   # This is not the best place to put this but before_test didn't work.
