+++ .travis.yml
@@ -23,7 +23,7 @@ install:
 
 
   # Now install eniric
-  - python setup.py install
+  - pip install .[dev,test,docs,ci]
 
 before_script:
   - export MPLBACKEND="AGG"  # Non-interactive
