### 2019-01-01 9286f51
fix manifest scripts
