### 2019-01-01 48fb0ee
Fix figure on home page
