### 2019-01-01 0ba1b77
Update installation procedure with new Starfish api.
Former-commit-id: d3dcbded55a0fdbee813209ff6add9f7ef2605e3 [formerly e395ec9f693ffef2ceffbed10162087c2781bdae]
Former-commit-id: 745f5ee29c78892edb7bbf5c92e1f7fcc511f344
