### 2019-01-01 0475ab9
adjust broadening
Former-commit-id: 26553e4b96fbf7c62e6e8fcbf928424f6f923704 [formerly d1b9704d93ff458d60b0e4713fd8967ffdb5ce6f]
Former-commit-id: d8ffdb4bc0d9ddd97c3498c432192fde89e85f45
