### 2019-01-01 087d0f7
Add rst formatting checks to travis-ci.
Former-commit-id: 2e72c072ff0e74cecaf662143370733f6c166767 [formerly 2c6c8a00435402fef4d90a6c048e94878cb87ab7]
Former-commit-id: e522cb59c8925b4a80a12a4f5f519366bd6991f4
