+++ setup.py
@@ -51,7 +51,8 @@ config = {
                  "hypothesis",
                  ],
         "docs": ["sphinx >= 1.4",
-                 "sphinx_rtd_theme"],
+                 "sphinx_rtd_theme",
+                 "rstcheck"],
         "ci": ["codacy-coverage==1.3.11",
                "codeclimate-test-reporter>=0.2.3",
                "python-coveralls>=2.9.1",
