+++ .travis.yml
@@ -43,6 +43,8 @@ script:
     - pytest --cov=eniric --cov-report term-missing --durations 5
     - coverage xml -i
 
+    - rstcheck -r docs
+
 after_success:
   - coveralls
   - codeclimate-test-reporter
