### 2019-01-01 6e74a07
Pin matplotlib to latest version 3.0.2
