+++ requirements.txt
@@ -1,6 +1,6 @@
 astropy>=3.0.0
 joblib>=0.12.3  # >=0.12.3
-matplotlib
+matplotlib==3.0.2
 multiprocess
 numpy>=1.15.4  # numpy>=1.15.4
 pandas
