### 2019-01-01 468bf12
Pin scipy to latest version 1.2.0
