+++ requirements.txt
@@ -5,7 +5,7 @@ multiprocess==0.70.6.1
 numpy>=1.15.4  # numpy>=1.15.4
 pandas==0.23.4
 pyyaml==3.13
-scipy
+scipy==1.2.0
 tqdm
 # Starfish install- currently needs numpy and cython pre-installed
 # https://github.com/iancze/Starfish/archive/master.zip#egg=Starfish
