### 2019-01-01 5c64296
Pin tqdm to latest version 4.28.1
