### 2019-01-02 2b62c50
Update requirements.txt
Remove added pins
