+++ requirements.txt
@@ -1,11 +1,11 @@
 astropy>=3.0.0
 joblib>=0.12.3  # >=0.12.3
-matplotlib==3.0.2
-multiprocess==0.70.6.1
+matplotlib
+multiprocess
 numpy>=1.15.4  # numpy>=1.15.4
-pandas==0.23.4
-pyyaml==3.13
-scipy==1.2.0
-tqdm==4.28.1
+pandas
+pyyaml
+scipy
+tqdm
 # Starfish install- currently needs numpy and cython pre-installed
 # https://github.com/iancze/Starfish/archive/master.zip#egg=Starfish
