+++ docs/source/Basic-usage.rst
@@ -4,27 +4,29 @@ Basic Usage
 
 To calculate the RV precision given a spectrum with ``wavelength`` and ``flux`` you can use the ``rv_precsion`` function
 
-.. code-block::
+.. autofunction eniric.Qcalculator :: rv_precision
+
+.. code-block:: python
 
    from eniric.Qcalculator import rv_precision
    rv = rv_precision(wavelength, flux)
 
 or for the flux independent spectral ``quality``
 
-.. code-block::
+.. code-block:: python
 
    from eniric.Qcalculator import quality
    q = quality(wavelength, flux)
 
 To apply a mask to the optimal pixel weights pass it as the 3rd argument.
 
-.. code-block::
+.. code-block:: python
 
    rv = rv_precision(wavelength, flux, my_mask)
 
-or 
+or
 
-.. code-block::
+.. code-block:: python
 
    rv = rv_precision(wavelength, flux, mask=my_mask)
 
@@ -32,14 +34,14 @@ Not suppling a mask or setting ``mask=None`` is equivalent to a mask of all ``1`
 
 To use the default telluric line model to apply masking:
 
-.. code-block::
+.. code-block:: python
 
    from eniric.atmosphere import Atmosphere
    # Assuming K is the correct band, and you want to mask seasonal variation.
-   atm = Atmosphere.from_band("K, bary=True)
+   atm = Atmosphere.from_band("K", bary=True)
 
    # Obtain closet telluric model values at the wavelength values (telluric mask is super sampled).
-   atm = atm.at(wavelength)  
+   atm = atm.at(wavelength)
 
    # Boolean telluric line mask at 2% deep.
    rv2 = rv_precision(wavelength, flux, mask=atm.mask)
@@ -47,18 +49,20 @@ To use the default telluric line model to apply masking:
    # Perfect telluric correction mask.
    rv3 = rv_precision(wavelength, flux, mask=atm.transmission**2)
 
-...
 
-As the presence of absorption lines in Earth's atmosphere blocks affects the spectral precision.
+The presence of absorption lines in Earth's atmosphere affects the spectral precision.
 There are two options to handle this. Complete masking of telluric lines within 30 km/s, reduction in spectral variance by the transmission spectrum.
 
 Weight Masking
 ~~~~~~~~~~~~~~
 
-Three cases: 
+Three cases:
 
 
 * all 1's
 * 1s and 0s
 * Transmission spectrum
-  The masking function is squared which only affects the 3rd option, If you want to alter the weigths in a specific way you will need to account for this when deriving the mask.
+  The masking function is squared which only affects the 3rd option, If you want to alter the weights in a specific way you will need to account for this when deriving the mask.
+
+
+Explanation of output file
