+++ docs/source/index.rst
@@ -44,11 +44,10 @@ Welcome to Eniric's documentation!
 
 
 `Eniric` is a Python 3.6+ software written to calculate the theoretical Radial Velocity (RV) precision of Near-InfraRed (NIR) stellar spectra.
-`Eniric` is an overhaul and extension to the code initially used in `Figueria et al. 2016 <http://dx.doi.org/10.1051/0004-6361/201526900>`_ to analysis the precision of M-dwarf stars in the NIR, extending its ability to use any synthetic spectra from the PHOENIX-ACES and BT-Settl libraries, or user provided spectra, and making it easier/faster to use.
-Extending the performance and usability, it is able to be used on any synthetic spectra from the `PHOENIX-ACES <http://phoenix.astro.physik.uni-goettingen.de>`_ and `BT-Settl <https://phoenix.ens-lyon.fr/Grids/BT-Settl/CIFIST2011_2015/FITS/>`_ (CIFIST2001-2015) libraries.
+`Eniric` is an overhaul and extension to the code initially used in `Figueira et al. 2016`_ to analysis the precision of M-dwarf stars in the NIR, extending its ability to use any synthetic spectra from the PHOENIX-ACES and BT-Settl libraries, or user provided spectra, and making it easier/faster to use.
+Extending the performance and usability, it is able to be used on any synthetic spectra from the `PHOENIX-ACES`_ and `BT-Settl`_ (CIFIST2001-2015) libraries.
 
-To get started see the `installation instructions <https://github.com/jason-neal/eniric/wiki/Installation>`_\ , or the `basic usage <https://github.com/jason-neal/eniric/wiki/Basic-usage>`_.
-
+To get started see the :doc:`Installation <./Installation>`, or the :doc:`Basic Usage <./Basic-usage>`.
 
 Features
 --------
@@ -58,7 +57,7 @@ Features
 * `Spectral broadening <https://github.com/jason-neal/eniric/wiki/Broadening>`_\ :
    Allows for Rotational and Instrumental broadening of synthetic spectra given rotation speed ``vsini`` and resolution ``R``.
 * `Atmospheric transmission <https://github.com/jason-neal/eniric/wiki/Atmospheric-Transmission>`_ masking:
-    Analyzing the RV precision attainable under the different masking conditions presented in `Figueria et al. 2016 <http://dx.doi.org/10.1051/0004-6361/201526900>`_.
+    Analyzing the RV precision attainable under the different masking conditions presented in `Figueira et al. 2016`_.
 
   * No treatment of atmospheric transmission
   * Masking all regions affected by atmospheric absorption of a given % over the course of the year.
@@ -66,7 +65,7 @@ Features
 
 * Relative RV precision
    The RV precision are present relative to a specified SNR per pixel in the center of a photometric band.
-   The default as used in the Figueira et al. 2016 is a SNR of 100 at the center of the J-band.
+   The default as used in the `Figueira et al. 2016`_ is a SNR of 100 at the center of the J-band.
 * Spectral Re-sampling
    Allows for re-sampling of spectra to `N` pixels per FWHM. Default = 3.
 * Photometric Band selection
@@ -78,8 +77,8 @@ Features
     Determine the incremental spectral quality or RV precision on narrow wavelength slices across the entire spectrum, similar to that present in Figure 1 of `Artigau et al. 2018 <http://adsabs.harvard.edu/abs/2018AJ....155..198A>`_.
 * Analyse precision of synthetic libraries
[…]
