### 2019-01-02 2218dbf
Add table describing output file columns.
Fixes #119
