### 2019-01-02 d73ebdc
Fix manifest.in - include
Former-commit-id: edb5afaf23b5cfbc4c84391e1448c4cde9397591 [formerly 55fd041aa1e39769fd77638d343bb70ec776a8e1]
Former-commit-id: 8f53d260a5be55b7fce1803ca0d9af4ed553808e
