### 2019-01-02 6b97984
Add eniric_suitable Starfish github repo into requirements.
Should enable automatic building now.
