### 2019-01-02 4d94270
Simplify Starfish install in docs and readme.
