### 2019-01-02 64f9d57
Simplify Starfish install in docs and readme.
Former-commit-id: b6d6f174a5460c993d59cefb6b9be582bc4f3b59 [formerly 6cb1fcfbadffe73b3f949fcb7110b8104dbe18a3]
Former-commit-id: fa4e4e12f56ad86f0dac1bbb58eacb91afc12ebe
