### 2019-01-03 e6974e3
Only put astrostarfish in requirements instead of https url for `install_requires`
