### 2019-01-03 043c7a5
Update appveyor.yml
Remove cd call
