+++ appveyor.yml
@@ -16,7 +16,6 @@ install:
   - "%PYTHON%\\python.exe -m pip install -r requirements.txt"
 
   # Installing eniric now
-  - cmd: "cd c:\\projects\\eniric"
   - "%PYTHON%\\python.exe setup.py install"
 
   # Setup config and make data
