### 2019-01-03 5eb73aa
Fix test_config paths for independent platforms path strings.
