+++ tests/test_broadening.py
@@ -78,7 +78,7 @@ def test_convolution_can_accept_int(num_proc):
     n = 20
     x = np.linspace(2.0, 2.3, n)
     y = np.random.randn(n)
-    convolution(x, y, vsini=1, R=100000, band="K", num_procs=num_proc)
+    convolution(x, y, vsini=1, R=100_000, band="K", num_procs=num_proc)
 
 
 @pytest.mark.parametrize("num_proc", [1, 2])
@@ -94,7 +94,7 @@ def test_res_convolution_can_accept_int(num_proc):
     n = 20
     x = np.linspace(2.0, 2.3, n)
     y = np.random.randn(n)
-    resolution_convolution(x, x, y, R=100000, num_procs=num_proc)
+    resolution_convolution(x, x, y, R=100_000, num_procs=num_proc)
 
 
 @pytest.mark.parametrize("num_proc", [1, 2])
@@ -103,7 +103,7 @@ def test_convolution_can_accept_worker_pool(num_proc):
     x = np.linspace(2.0, 2.3, n)
     y = np.random.randn(n)
     with Pool(num_proc) as mproc:
-        convolution(x, y, vsini=1, R=100000, band="K", num_procs=mproc, verbose=False)
+        convolution(x, y, vsini=1, R=100_000, band="K", num_procs=mproc, verbose=False)
 
 
 @pytest.mark.parametrize("num_proc", [1, 2])
@@ -121,7 +121,7 @@ def test_res_convolution_can_accept_worker_pool(num_proc):
     x = np.linspace(2.0, 2.3, n)
     y = np.random.randn(n)
     with Pool(num_proc) as mproc:
-        resolution_convolution(x, x, y, R=100000, num_procs=mproc, verbose=False)
+        resolution_convolution(x, x, y, R=100_000, num_procs=mproc, verbose=False)
 
 
 @pytest.mark.parametrize("num_proc", [3.14, "str"])
@@ -143,14 +143,14 @@ def test_res_convolution_with_bad_num_proc(num_proc):
     with pytest.raises(
         TypeError, message="num_proc must be an int or a multiprocess Pool"
     ):
-        resolution_convolution(x, x, y, R=100000, num_procs=num_proc, verbose=False)
+        resolution_convolution(x, x, y, R=100_000, num_procs=num_proc, verbose=False)
 
 
 def test_convolution_can_accept_None():
     n = 20
     x = np.linspace(2.0, 2.3, n)
     y = np.random.randn(n)
-    convolution(x, y, vsini=1, R=100000, band="K", num_procs=None)
+    convolution(x, y, vsini=1, R=100_000, band="K", num_procs=None)
 
 
 def test_rot_convolution_can_accept_None():
@@ -164,7 +164,7 @@ def test_res_convolution_can_accept_None():
     n = 20
     x = np.linspace(2.0, 2.3, n)
     y = np.random.randn(n)
-    resolution_convolution(x, x, y, R=100000, num_procs=None)
+    resolution_convolution(x, x, y, R=100_000, num_procs=None)
 
 
 @pytest.mark.parametrize("zero_vsini", [0, 0.00, np.int(0.0), np.float64(0.0)])
