### 2019-01-03 7cf2204
Remove obsolete modules, function, config and script calls.
