+++ eniric/obsolete/utilities.py
@@ -1,295 +1,0 @@
-import os
-import re
-from genericpath import isfile
-from os.path import join
-from typing import Optional, Tuple, Union
-
-import numpy as np
-from astropy import constants as const
-from numpy.core.multiarray import ndarray
-
-import eniric
-import eniric.obsolete.IOmodule
-from eniric import io_module as io
-from eniric.atmosphere import consecutive_truths
-from eniric.resample import log_resample
-
-
-def get_spectrum_name(
-    startype: str,
-    logg: Union[float, int] = 4.50,
-    feh: Union[float, int] = 0.0,
-    alpha: Optional[Union[int, float]] = None,
-    org: bool = False,
-    flux_type: str = "photon",
-) -> str:
-    """Return correct phoenix spectrum filename for a given spectral type.
-
-    Based off phoenix_utils module.
-
-    Parameters
-    ----------
-    flux_type: str
-        Indicate which file type to try find. e.g. "photon", "wave", ("fits" Not Implemented yet)
-
-    Returns
-    -------
-    spectrum_name: str
-        The name of spectrum with chosen Parameters
-
-
-    Ability to select logg and metallicity (feh) later on.
-    org = original locations (without Z folder option)
-    """
-    if feh == 0:
-        feh = -0.0  # make zero negative to signed integer.
-
-    temps = {"M0": 3900, "M3": 3500, "M6": 2800, "M9": 2600}
-    if (flux_type == "photon") and (not org):
-        base = "PHOENIX-ACES-AGSS-COND-2011-HiRes_wave_photon.dat"
-    else:
-        base = "PHOENIX-ACES-AGSS-COND-2011-HiRes_wave.dat"
-
-    # noinspection SpellCheckingInspection
-    if startype in temps.keys():
-        if org:
-            phoenix_name = "lte{0:05d}-{1}-{2}.{3}".format(
-                temps[startype], "4.50", "0.0", base
-            )
-        elif (alpha is not None) and (alpha != 0.0):
-            if abs(alpha) > 0.2:
-                raise ValueError(
-                    "Warning! Alpha is outside acceptable range -0.2->0.2. (for current science case)"
-                )
-
-            phoenix_name = os.path.join(
-                "Z{0:+4.1f}.Alpha={1:+5.2f}".format(feh, alpha),
-                "lte{0:05d}-{1:4.2f}{2:+4.1f}.Alpha={3:+5.2f}.{4:s}".format(
-                    temps[startype], logg, feh, alpha, base
-                ),
-            )
-        else:
-            phoenix_name = os.path.join(
-                "Z{0:+4.1f}".format(feh),
-                "lte{0:05d}-{1:4.2f}{2:+4.1f}.{3:s}".format(
-                    temps[startype], logg, feh, base
-                ),
-            )
-
-        spectrum_name = phoenix_name
-    elif re.match(r"^[OBAFGKML][0-9]$", startype):  # Valid spectral types
-        raise NotImplementedError(
-            "The spectral type '{0:s}' is not implemented yet.".format(startype)
-        )
-    else:
-        raise ValueError("'{0:s}' is not a valid spectral type.".format(startype))
-
-    return spectrum_name
-
-
-def barycenter_shift(
-    wav_atm: ndarray,
-    mask_atm: ndarray,
-    rv_offset: float = 0.0,
-    consecutive_test: bool = True,
-) -> ndarray:
-    """Calculating impact of Barycentric movement on mask...
-
-    Extends the masked region to +-30 km/s due to the barycentric motion of the earth.
-    """
-    # Mask values to the left and right side of mask_atm.
-    # To avoid indexing errors have padded with first and last values.
-    mask_iminus1 = np.concatenate(
-        ([mask_atm[0], mask_atm[0]], mask_atm[:-2])
[…]
