### 2019-01-03 50a7084
Pip install Strfish/astrostarfish (#126)
* Pip install starfish after starfish change.

* Add eniric_suitable Starfish github repo into requirements.

Should enable automatic building now.

* Only put astrostarfish in requirements instead of https url for `install_requires`

* Fix test_config paths for independent platforms path strings.

* Reformat with Black.
