+++ eniric/resample.py
@@ -71,7 +71,7 @@ def log_chunks(wavelength, percent):
     Allows spectrum to be split into chunks in which the size is X% of the given wavelength.
     This takes logarithmic steps with a base of (1+X/100).
     """
-    base = 1 + percent / 100.
+    base = 1 + percent / 100.0
     wl_min = np.nanmin(wavelength)
     wl_max = np.nanmax(wavelength)
 
