+++ .travis.yml
@@ -15,12 +15,6 @@ install:
   - travis_retry pip install -r requirements_dev.txt
   - travis_retry pip install -r requirements.txt
 
-  # Install Starfish
-  - wget -O starfishrequirements.txt https://raw.githubusercontent.com/iancze/Starfish/master/requirements.txt
-  - pip install -r starfishrequirements.txt
-  # - pip install https://github.com/iancze/Starfish/archive/master.zip#egg=Starfish
-  - pip install -e git+https://github.com/iancze/Starfish.git#egg=Starfish  # Editable install that includes build_ext
-
 
   # Now install eniric
   - python setup.py install
