### 2019-01-03 3546531
Pip install Strfish/astrostarfish (#126)
* Pip install starfish after starfish change.

* Add eniric_suitable Starfish github repo into requirements.

Should enable automatic building now.

* Only put astrostarfish in requirements instead of https url for `install_requires`

* Fix test_config paths for independent platforms path strings.

* Reformat with Black.


Former-commit-id: 3a498e967a3ae902638b6efbe208d46e8c2e9a5a [formerly e38ef8f0cc74c934c242d2c1a91d9b084e41b92d]
Former-commit-id: d321c4f983282fca4709d4b38a95df95f29c5fc7
