### 2019-01-03 ff05220
Fail rstcheck on errors only.
Former-commit-id: 5aee489db362da8ed405f8abfbedb5d3a5f131cf [formerly c336f6ae3d81a48b9445d9e2acd33bf8f8394e4a]
Former-commit-id: 35961888f52ce92b63892b2d0955b5af6cfc2ab2
