### 2019-01-03 209da7b
Fail rstcheck on errors only.
