### 2019-01-13 #1053 @jason-neal
I have achieved some progress with fixing up the documentation and putting it on read the docs, but it is still a work in progress https://eniric.readthedocs.io/en/newdocs/.

I am currently focusing on a hard thesis deadline of February 8th.
