### 2019-01-29 2d6d263
Refactor bary_shift_mask to barycenter_broaden
Former-commit-id: 30da5cedb14f5ce00d75c5369f86430623e265b7 [formerly 7af6854b83e1b186f264ba136ddf362a25201f69]
Former-commit-id: 5e44047d7871698884994704b2ec31856755a1f3
