### 2019-01-29 671942a
Refactor bary_shift_mask to barycenter_broaden
