### 2019-02-01 1843d61
Pin matplotlib to latest version 3.0.2
