### 2019-02-01 a5fa6e9
Pin pandas to latest version 0.24.0
