+++ requirements.txt
@@ -4,7 +4,7 @@ joblib>=0.12.3  # >=0.12.3
 matplotlib==3.0.2
 multiprocess==0.70.7
 numpy>=1.15.4  # numpy>=1.15.4
-pandas
+pandas==0.24.0
 pyyaml
 scipy
 tqdm
