### 2019-02-01 1b5b7dd
Pin pyyaml to latest version 3.13
