+++ requirements.txt
@@ -5,6 +5,6 @@ matplotlib==3.0.2
 multiprocess==0.70.7
 numpy>=1.15.4  # numpy>=1.15.4
 pandas==0.24.0
-pyyaml
+pyyaml==3.13
 scipy
 tqdm
