### 2019-02-01 06e4294
Pin scipy to latest version 1.2.0
