### 2019-02-01 8a27d8e
Pin tqdm to latest version 4.30.0
