### 2019-02-18 #1053 @jason-neal
The thesis is now submitted so I can focus on completing this.
I have made quite a few of the changes (with commits referencing the issues) in the [newdocs](https://github.com/jason-neal/eniric/tree/newdocs) branch.

I still have some issues with learning/getting the auto-documentation correct on readthedocs.
