### 2019-02-18 e8824ce
Scheduled monthly dependency update for February (#127)
* Pin matplotlib to latest version 3.0.2

* Pin multiprocess to latest version 0.70.7

* Pin pandas to latest version 0.24.0

* Pin pyyaml to latest version 3.13

* Pin scipy to latest version 1.2.0

* Pin tqdm to latest version 4.30.0

* Update hypothesis from 3.85.2 to 4.4.3

* Update joblib from 0.13 to 0.13.1

* Update multiprocess from 0.70.6.1 to 0.70.7

* Update mypy from 0.650 to 0.660

* Update numpy from 1.15.4 to 1.16.1

* Update pandas from 0.23.4 to 0.24.0

* Update pre-commit from 1.13.0 to 1.14.2

* Update pytest from 4.0.2
[…]
