+++ requirements.txt
@@ -1,10 +1,10 @@
 astropy>=3.0.0
 astrostarfish
 joblib>=0.12.3  # >=0.12.3
-matplotlib
-multiprocess
+matplotlib==3.0.2
+multiprocess==0.70.7
 numpy>=1.15.4  # numpy>=1.15.4
-pandas
-pyyaml
-scipy
-tqdm
+pandas==0.24.0
+pyyaml==3.13
+scipy==1.2.0
+tqdm==4.30.0
