### 2019-02-19 cfa5009
Add Citations to packages in paper.
Fixes #111
