### 2019-02-19 f0e60ce
Move download scripts to scripts/download.
Create new download script using new Starfish functions.

Cache data on CI services
