+++ setup.py
@@ -49,8 +49,9 @@ config = {
         "scripts/split_atmmodel.py",
         "scripts/bary_shift_atmmodel.py",
         "scripts/untar_here.py",
-        "scripts/download_eniric_data.sh",
-        "tests/download_test_PHOENIX_spec.sh",
+        "scripts/download/download_eniric_data.sh",
+        "scripts/download/download_test_PHOENIX_spec.sh",
+        "scripts/download/download_test_aces.py",
         "eniric/obsolete/make_test_data.py",
         "eniric/obsolete/nIR_run.py",
         "eniric/obsolete/nIR_precision.py",
