+++ appveyor.yml
@@ -7,6 +7,10 @@ environment:
     - PYTHON: "C:\\Python36-x64"
     - PYTHON: "C:\\Python37-x64"
 
+cache:
+  - "%LOCALAPPDATA%\pip\Cache -> appveyor.yml" # Use cache unless appveyor.yml changes.
+  - "data\test_data -> appveyor.yml"           # Use cache unless appveyor.yml changes.
+
 install:
   - "%PYTHON%\\python.exe -m pip install --upgrade pip"
   # We need wheel installed to build wheels
