+++ requirements.txt
@@ -1,10 +1,10 @@
 astropy>=3.0.0
 astrostarfish
 joblib>=0.12.3  # >=0.12.3
-matplotlib==3.0.2
-multiprocess==0.70.7
+matplotlib
+multiprocess
 numpy>=1.15.4  # numpy>=1.15.4
-pandas==0.24.0
-pyyaml==3.13
-scipy==1.2.0
-tqdm==4.30.0
+pandas
+pyyaml
+scipy
+tqdm
