### 2019-02-20 e4c0669
Fix download_test_aces.py Starfish downloaded takes a iterable of iterables.
