### 2019-02-20 jason-neal/eniric#111 @jason-neal
References to the packages have been added as well as hyperlinks to the webpages.
