### 2019-02-20 7066210
Depreciate the model name ´phoenix´.
Added a model check with tests.

PR #128
