### 2019-02-20 10863e6
fix atmos make command
Former-commit-id: 45643c2b38186d1b5faa33fbb28b11e0905af0d4 [formerly 3d644d9b1ece9816227dc84a2978ba2060ca63d5]
Former-commit-id: 597e2c16b5e49ff61182305e596c4e5f5b576678
