+++ Makefile
@@ -47,7 +47,7 @@ data:
 
 atmos:
 	split_atmmodel.py -b ALL
-	baryc_shift_atmmodel.py -b ALL
+	barycenter_broaden_atmmodel.py -b ALL
 
 clean-build:
 	rm --force --recursive build/
