### 2019-02-20 9da31ce
Fix Atmosphere.barycenter_broaden() method calls
