### 2019-02-20 689783c
Merge some setting changes from newdocs branch.
Former-commit-id: 66eaa17c4e5af020d55fa38a1d512f431e942b59 [formerly 5bb2a3f3fd249a0c732d1b0e5b515b5c4f2a8656]
Former-commit-id: 82ab90ad280490a7d6d998972ff87f07d3882def
