### 2019-02-20 2e6f01b
Fix comma bug in example usage.
Fixes #116.
