### 2019-02-20 jason-neal/eniric#112 @jason-neal
I have also added a link to https://eniric.readthedocs.io/en/latest/ in the Github page description/website section.
