### 2019-02-20 d2c7be4
Use pytest.deprecated_call
