### 2019-02-21 8fda28c
Small tweaks, mainly codacy issues.
Former-commit-id: 028883e283a060294dae503087fdea1742c82226 [formerly a31550a40b54866c1ef353fbec1fc59f3655d71c]
Former-commit-id: 6f7c2d5ea4d11f23deafaf559f4519f4fc4e4f9f
