### 2019-02-21 2d2928d
Small tweaks, mainly codacy issues.
