### 2019-02-21 29352ad
Fix appveyor escape characters in paths \\.
