### 2019-02-21 0991d4d
Change --model flag options (#130)
* Depreciate the model name ´phoenix´.

Added a model check with tests.

PR #128

* Use pytest.deprecated_call

* Update changelog
