### 2019-02-21 e7f044f
Change --model flag options (#130)
* Depreciate the model name ´phoenix´.

Added a model check with tests.

PR #128

* Use pytest.deprecated_call

* Update changelog


Former-commit-id: e97aae23b1becac60aacdef8f55c4a7c0e0dabb5 [formerly 87693f4ff7a73db18dbe23a95d3bd9be4e67f941]
Former-commit-id: 4d9e19f514b4960dad27cd07ad32dadf5dae5309
