### 2019-02-22 2e492d9
Update doc strings.
Specifically adjust Default values.


Former-commit-id: ea117797c2651d61ddd4780331c5e6436d7ef7f6 [formerly 3c98fd6078e6af1966940e82e53ee06015765c6a]
Former-commit-id: 4218dc7eb754635c50647e20f46411413d2374e7
