### 2019-02-22 d59cf57
Update doc strings.
Specifically adjust Default values.
