+++ eniric/broaden.py
@@ -1,8 +1,8 @@
-"""Broadening functions
+"""Spectral broadening functions:
 
-Used to convolve the spectra for
-    - Stellar rotation
-    - Instrumental resolution
+Used to convolve the model spectra for
+    - Stellar rotation broadening.
+    - Instrumental profile broadening.
 
 Uses joblib.Memory to cache convolution results to skip repeated computation.
 """
@@ -39,54 +39,54 @@ def rotational_convolution(
     num_procs: Optional[Union[int, pool.Pool]] = None,
     verbose: bool = True,
 ) -> ndarray:
-    """Perform Rotational convolution.
+    r"""Perform Rotational convolution.
 
     Parameters
     ----------
     wavelength: ndarray
-        Wavelength to calculate convolution at.
+        Wavelength array.
     extended_wav: ndarray
-       Wavelength extended to avoid boundary issues.
+       Extended wavelength array to avoid boundary issues.
     extended_flux: ndarray
-        Photon flux, at extend wavelength
+        Photon flux for the extended wavelength array.
     vsini: float
         Rotational velocity in km/s.
-    epsilon: float (default = 0.6)
-        Limb darkening coefficient
-    normalize: bool (default = True)
-        Area normalize the broadening kernel (corrects for unequal sampling of position).
+    epsilon: float
+        Limb darkening coefficient. Default is 0.6.
+    normalize: bool
+        Area normalize the broadening kernel. This corrects for kernel area with unequal wavelength spacing.
+        Default is True.
     num_procs: int, None or multiprocess.pool.Pool.
-        Number of processes to use with multiprocess.
-        If None it is assigned to 1 less then total number of cores.
+        Number of processes to use.
+        If None then 1 less then total number of cores is used.
         If num_procs = 0 or 1, then multiprocess is not used.
         Can also be a multiprocess.pool.Pool instance.
     verbose: bool
-        Show the tqdm progress bar (default = True).
+        Show the tqdm progress bar. Default is True.
 
     Returns
     -------
     convolved_flux: ndarray
-        The convolved_flux evaluated at wavelength points.
+        The convolved flux evaluated at the wavelength array.
 
     """
 
     def element_rot_convolution(single_wav: float) -> float:
-        """Embarrassingly parallel part of rotational convolution.
+        r"""Embarrassingly parallel part of rotational convolution.
 
         Calculates the convolution value for a single pixel.
-
         The parameters extended_wav, extended_flux, vsini, epsilon and normalize
         are obtained from the outer scope.
 
         Parameters
         ----------
         single_wav: float
-            Wavelength value to calculate convolution at.
+            Wavelength to calculate the convolution at.
 
         Returns
         -------
         sum_val: float
-            Sum of flux convolved for this pixel/wavelength.
+            Sum of flux convolved for this wavelength.
 
         """
         # Select all values such that they are within the fwhm limits
@@ -164,38 +164,39 @@ def resolution_convolution(
     num_procs: Optional[Union[int, pool.Pool]] = None,
     verbose: bool = True,
 ) -> ndarray:
-    """Perform Resolution convolution.
+    r"""Perform Resolution convolution.
 
     Parameters
     ----------
     wavelength: ndarray
-        Wavelength in microns to
+        Wavelength array.
     extended_wav: ndarray
-        Wavelength array slightly longer to avoid boundary errors.
+       Extended wavelength array to avoid boundary issues.
     extended_flux: ndarray
-        Photon flux
[…]
