### 2019-02-22 3e9cabc
Update broadening.rst
Former-commit-id: 764e67e60884a89b57657591c6b42223494443b5 [formerly 345f2fc016c989be9086b7f1421b6d4d988c99f7]
Former-commit-id: 0568cd5051865c4cd062b40acbd300335230ed10
