### 2019-03-01 28b4620
Pin matplotlib to latest version 3.0.3
