### 2019-03-01 294f97f
Pin pandas to latest version 0.24.1
