### 2019-03-01 8702886
Pin pyyaml to latest version 3.13
