### 2019-03-01 ec7c314
Pin scipy to latest version 1.2.1
