### 2019-03-01 3d8d913
Pin tqdm to latest version 4.31.1
