+++ requirements.txt
@@ -7,4 +7,4 @@ numpy>=1.15.4  # numpy>=1.15.4
 pandas==0.24.1
 pyyaml==3.13
 scipy==1.2.1
-tqdm
+tqdm==4.31.1
