### 2019-03-04 d1ae08d
Paper updates (#129)
* Edits to JOSS paper following #110

* Add Citations to packages in paper.

Fixes #111


Former-commit-id: 5cbb59bb3543eb2ebf429ab306cd0b6a8324c423 [formerly 6d741390933f6652018336573ed2f96a4e13029c]
Former-commit-id: cc46f263fb9864dc6c8163120ef81f651d91dd39
