### 2019-03-04 0a28885
Paper updates (#129)
* Edits to JOSS paper following #110

* Add Citations to packages in paper.

Fixes #111
