### 2019-03-06 e7b84c9
Add default temperature value to argparse in script.
Tweak help messages.
Fixes #132.
