### 2019-03-06 88bf5e9
load_flux does not return header.
Former-commit-id: 495dab053b41588517a639594e66fb33da47df9d [formerly 32f6080b9a8b832fbb21dd3de81252a07b2551ef]
Former-commit-id: 28e262ee6d58fbe05796f85fe9a787f0f88f87bb
