+++ eniric/utilities.py
@@ -320,7 +320,7 @@ def load_aces_spectrum(
         raise ValueError("Number of parameters is incorrect")
 
     wav = phoenix_grid.wl
-    flux, __ = phoenix_grid.load_flux(params, norm=False)
+    flux = phoenix_grid.load_flux(params, norm=False)
 
     # Convert wavelength Angstrom to micron
     wav_micron = wav * 10 ** -4
@@ -396,7 +396,7 @@ def load_btsettl_spectrum(
 
     wav = btsettl_grid.wl
     # CIFIST flux is  W/m**2/um
-    flux, __ = btsettl_grid.load_flux(params, norm=False)
+    flux = btsettl_grid.load_flux(params, norm=False)
 
     # [::10] down samples only every 10th point from BT-Settl CIFIST spectrum.
     wav, flux = wav[::10], flux[::10]
