+++ docs/source/Broadening.rst
@@ -40,16 +40,13 @@ Gaussian kernel
 
 Circular Fibre
 ++++++++++++++
+This is the convolution kernel for a circular fibre.
 .. autofunction:: oned_circle_kernel
 
 When analyzing the spectral libraries, rotational broadening is preformed first, followed by the instrumental broadening.
 
 Our convolution functions use wavelength dependent kernels and do not require uniform spacing between points, unlike `PyAstronomy <https://www.hs.uni-hamburg.de/DE/Ins/Per/Czesla/PyA/PyA/pyaslDoc/aslDoc/broadening.html>`_.
 This means our convolutions are slower but are more precise. We compare the convolutions in the `Convolution_speeds.ipynb <https://github.com/jason-neal/eniric/blob/develop/docs/Notebooks/Convolution_speeds.ipynb>`_ notebook.
-
-Caching of the convolution stages is performed to avoid re-computation of this slow component when possible using `\ ``joblib.Memory``  <https://joblib.readthedocs.io/en/latest/memory.html>`_. The caching directory defaults to ``~/.joblib`` but can be changed in the configuration file ``config.yaml``.
-
-Caching can be disabled by setting ``location=None``
 
 
 Combined convolution
@@ -64,3 +61,7 @@ Caching
 
 Convolution results are cached using :mod:`joblib`, see https://joblib.readthedocs.io/en/latest/memory.html
 
+
+Caching of the convolution stages is performed to avoid re-computation of this slow component when possible using `\ ``joblib.Memory``  <https://joblib.readthedocs.io/en/latest/memory.html>`_. The caching directory defaults to ``~/.joblib`` but can be changed in the configuration file ``config.yaml``.
+
+Caching can be disabled by setting ``location=None`` in ``config.yml``.
