### 2019-03-07 1f4e1e3
Add -h flag output to documentation.
