### 2019-03-07 2bf78d9
Add -h flag output to documentation.
Former-commit-id: dc29f68ac71b0ba3c18161c77ab19a9f1fc4a7b3 [formerly 8e274d9bccad22299b3cf44a96f39bab36952d6d]
Former-commit-id: 49367917d33cbf41a3bf917d9333f7d1507e7c06
