+++ docs/source/Examples.rst
@@ -1,0 +1,36 @@
+Example Notebooks
+=================
+
+There are several `Jupyter Notebooks <https://github.com/jason-neal/eniric/blob/master/docs/Notebooks/>`_ which focus on different aspects of eniric.
+
+Below are the links to their location in github or rendered with `nbveiwer <https://nbviewer.jupyter.org/>`_.
+
+
+These are:
+
+
+* Atmosphere model (`Github <https://github.com/jason-neal/eniric/blob/master/docs/Notebooks/atmosphere_example.ipynb>`_) (`nbviwer <https://nbviewer.jupyter.org/github/jason-neal/eniric/blob/master/docs/Notebooks/Compare_precisions.ipynb>`_)
+* Compare precisions (`Github <https://github.com/jason-neal/eniric/blob/master/docs/Notebooks/Compare_precisions.ipynb>`_) (`nbviwer <https://nbviewer.jupyter.org/github/jason-neal/eniric/blob/master/docs/Notebooks/Compare_precisions.ipynb>`_)
+
+* Incremental_precision.ipynb (`Github <https://github.com/jason-neal/eniric/blob/master/docs/Notebooks/Incremental_precision.ipynb>`_) (`nbviwer <https://nbviewer.jupyter.org/github/jason-neal/eniric/blob/master/docs/Notebooks/Incremental_precision.ipynb>`_)
+* Precision_example.ipynb (`Github <https://github.com/jason-neal/eniric/blob/master/docs/Notebooks/Precision_example.ipynb>`_) (`nbviwer <https://nbviewer.jupyter.org/github/jason-neal/eniric/blob/master/docs/Notebooks/Precision_example.ipynb>`_)
+* Precison_with_doppler-K-band.ipynb (`Github <https://github.com/jason-neal/eniric/blob/master/docs/Notebooks/Precison_with_doppler-K-band.ipynb>`_) (`nbviwer <https://nbviewer.jupyter.org/github/jason-neal/eniric/blob/master/docs/Notebooks/Precison_with_doppler-K-band.ipynb>`_)
+* Precison_with_doppler-Z-band.ipynb (`Github <https://github.com/jason-neal/eniric/blob/master/docs/Notebooks/Precison_with_doppler-Z-band.ipynb>`_) (`nbviwer <https://nbviewer.jupyter.org/github/jason-neal/eniric/blob/master/docs/Notebooks/Precison_with_doppler-Z-band.ipynb>`_)
+* Solar_precision.ipynb (`Github <https://github.com/jason-neal/eniric/blob/master/docs/Notebooks/Solar_precision.ipynb>`_) (`nbviwer <https://nbviewer.jupyter.org/github/jason-neal/eniric/blob/master/docs/Notebooks/Solar_precision.ipynb>`_)
+* Effect of spectral gradient (`Github <https://github.com/jason-neal/eniric/blob/master/docs/Notebooks/spectral_gradient.ipynb>`_) (`nbviwer <https://nbviewer.jupyter.org/github/jason-neal/eniric/blob/master/docs/Notebooks/spectral_gradient.ipynb>`_)
+
+* Split_verse_Weighted_masking (`Github <https://github.com/jason-neal/eniric/blob/master/docs/Notebooks/Split_verse_Weighted_masking.ipynb>`_) (`nbviwer <https://nbviewer.jupyter.org/github/jason-neal/eniric/blob/master/docs/Notebooks/Split_verse_Weighted_masking.ipynb>`_)
+
+Misc
+^^^^
+
[…]
