### 2019-03-07 92a7329
Update Theoretical precision section.
Update and move output file name and ordering.
