+++ docs/source/Basic-usage.rst
@@ -63,41 +63,3 @@ Three cases:
 * 1s and 0s
 * Transmission spectrum
   The masking function is squared which only affects the 3rd option, If you want to alter the weights in a specific way you will need to account for this when deriving the mask.
-
-
-Explanation of output file
-==========================
-
-``phoenix_precision.py`` saves the results to a output file.
-The different columns of the output file are given in the table below.
-
-+--------+--------------+---------------------------------------------------------------------+
-| Col.   | Name         | Description                                                         |
-+========+==============+=====================================================================+
-| 1      | Temp         | Library stellar effective temperature (K)                           |
-+--------+--------------+---------------------------------------------------------------------+
-| 2      | logg         | Library stellar surface gravity                                     |
-+--------+--------------+---------------------------------------------------------------------+
-| 3      | [Fe/H]       | Library stellar metallicity                                         |
-+--------+--------------+---------------------------------------------------------------------+
-| 4      | Band         | Wavelength band used                                                |
-+--------+--------------+---------------------------------------------------------------------+
-| 5      | Resolution   | Instrumental resolution                                             |
-+--------+--------------+---------------------------------------------------------------------+
-| 6      | vsini        | Stellar rotation (km/s)                                             |
-+--------+--------------+---------------------------------------------------------------------+
-| 7      | sampling     | Spectral sampling - N points per resolution element.                |
-+--------+--------------+---------------------------------------------------------------------+
-| 8      | Quality      | Theoretical spectral quality                                        |
-+--------+--------------+---------------------------------------------------------------------+
-| 9      | Cond. 1      | RV precision with no masking.                                       |
-+--------+--------------+---------------------------------------------------------------------+
-| 10     | Cond. 2      | RV precision with binary masking.                                   |
-+--------+--------------+---------------------------------------------------------------------+
-| 11     | Cond. 3      | RV precision with transmission masking.                             |
-+--------+--------------+---------------------------------------------------------------------+
-| 12     | correct flag | Indicate if `Artigau et al. 2018`_ precision correction is applied. |
-+--------+--------------+---------------------------------------------------------------------+
-
-
-.. _`Artigau et al. 2018`: http://adsabs.harvard.edu/abs/2018AJ....155..
