+++ docs/source/Theoretical-Precision-of-PHONEIX-ACES-Spectra.rst
@@ -1,31 +1,30 @@
 
 Theoretical Precision of Synthetic Spectra
-------------------------------------------
+******************************************
 
-Eniric provides a script ``phoenix_precision.py`` to generate RV precision values for any spectra in the PHOENIX-ACES spectral library (Available at `http://phoenix.astro.physik.uni-goettingen.de/ <http://phoenix.astro.physik.uni-goettingen.de/>`_\ ). The precision of a spectra can be obtained by providing its library parameters Teff, logg, Fe/H and, alpha. This has been mainly used on M-dwarf spectra (< 4000K), but it higher temperatures should work to.
+The theoretical precision of M-dwarf synthetic spectra is extensively explored in `Figueira et al. (2016)`_.
+Eniric extends this work to any spectra in the `PHOENIX-ACES <http://phoenix.astro.physik.uni-goettingen.de/>`_ or the `BT-Settl-CIFIST2011_2015 <https://phoenix.ens-lyon.fr/Grids/BT-Settl/CIFIST2011_2015/>`_ spectral libraries.
 
-For the library selection and loading of the spectra we use `Starfish's Grid Tools <https://iancze.github.io/Starfish/current/grid_tools.html>`_. You need to have the PHOENIX-ACES spectra downloaded, and the ``raw_path`` configured in the ``config.yaml`` file in the working directory. For more information see the Starfish documentation `here <https://iancze.github.io/Starfish/current/grid_tools.html#downloading-model-spectra>`_.
+The script ``phoenix_precision.py`` is provided to easily generate synthetic RV precision values, similarly to and beyond the work of `Figueira et al. (2016)`_.
+The precision of a spectrum can be obtained by providing its library parameters Teff, logg, Fe/H and, alpha. This has been mainly used on M-dwarf spectra with temperatures < 4000K, but higher temperatures also work.
 
-After downloading the PHOENIX-ACES spectra spectra you need to, set the ``raw_path`` to the downloaded files in the ``config.yaml`` file in the working directory.   
+For the library selection and loading of the spectra `Starfish's Grid Tools <https://iancze.github.io/Starfish/current/grid_tools.html>`_ is used.
+The spectral libraries need to be pre-downloaded, and the ``raw_path`` to their location configured in the ``config.yaml`` file in the working directory.
+For more information see the Starfish documentation `here <https://iancze.github.io/Starfish/current/grid_tools.html#downloading-model-spectra>`_.
 
-The script returns a table with input parameters and calculated spectral quality and precisions for all three conditions presented in Figueira et al. 2016 for all parameter combinations requested.   
-
-::
-
-   # quality.csv
-   Temp, logg, [Fe/H], Alpha, Band, Resolution, vsini, Sampling, Quality, Cond. 1, Cond. 2, Cond. 3, correct flag
-   3900, 4.5,     0.0,   0.0,    J,       100k,   1.0,      3.0,    3501,     7.6,    20.2,     8.0,            0
 
 Example Usage
-^^^^^^^^^^^^^
+=============
 
-You use this script by passing it stellar parameters for the spectral library as well as broadening and wavelength parameters.
-Multiple parameters can be passed to each flag, separated by a space.
-To see all the available inputs parameters run:
+``phoenix_precision.py`` is called by passing the stellar and other parameters on the command line.
+Several flags can take multiple values separated by a space. All combinations of the parameters supplied will be computed.
+
+The available inputs parameters are:
 
 .. code-block:: bash
 
     $ phoenix_precision.py -h
+
     usage: phoenix_precision.py [-h] [-t TEMP [TEMP ...]] [-l LOGG [LOGG ...]]
[…]
