+++ docs/source/Theoretical-Precision-of-PHONEIX-ACES-Spectra.rst
@@ -104,7 +104,7 @@ The different columns of the output file are given in the table below.
 +--------+--------------+----------------------------------------------------------------------+
 | 4      | alpha        | Library stellar alpha ratio.                                         |
 +--------+--------------+----------------------------------------------------------------------+
-| 5      | band         | Wavelength band used.                                                |
+| 5      | band         | Wavelength band letter.                                              |
 +--------+--------------+----------------------------------------------------------------------+
 | 6      | resolution   | Instrumental resolution.                                             |
 +--------+--------------+----------------------------------------------------------------------+
