### 2019-03-08 0517a7a
Refactor Qcalculator -> precision
