+++ eniric/__init__.py
@@ -25,7 +25,7 @@ __all__ = [
     "corrections",
     "legacy",
     "io_module",
-    "Qcalculator",
+    "precision",
     "resample",
     "snr_normalization",
     "utilities",
