+++ eniric/config.yaml
@@ -9,7 +9,7 @@ paths:
   test_data: ["..", "data", "test_data"]
   atmmodel: ["..", "data", "atmmodel"]
 
-  precision: ["precision"]  # A place to put precision results
+  precision_results: ["precision"]  # A place to put precision results
   # Obsolete paths for testing only
   phoenix_dat: ["data", "test_data", "obsolete", "PHOENIX-ACES_spectra"]
   results:  ["..", "data", "test_data", "obsolete", "results"]
