### 2019-03-08 efbe5d4
Refactor Qcalculator -> precision
Former-commit-id: 990a87f1306244cc88749ea59e716769d9695da5 [formerly 3fc9191fa8897a9036b55ff608eff552ea19a5b7]
Former-commit-id: 0f6929ffeaf529af8c674e055d46428240000707
