+++ tests/test_precision.py
@@ -1,4 +1,4 @@
-"""Test Qcalculator."""
+"""Test eniric.precision."""
 
 import astropy.units as u
 import numpy as np
@@ -6,9 +6,16 @@ import pytest
 from astropy import constants as const
 from astropy.units import Quantity
 
-import eniric.Qcalculator as Q
 from eniric.atmosphere import Atmosphere
-from eniric.Qcalculator import mask_check, pixel_weights
+from eniric.precision import (
+    incremental_quality,
+    incremental_rv,
+    mask_check,
+    pixel_weights,
+    quality,
+    rv_precision,
+    sqrt_sum_wis,
+)
 from eniric.snr_normalization import snr_constant_band
 from eniric.utilities import (
     band_limits,
@@ -32,7 +39,7 @@ def test_rvprev_calc(test_spec, wav_unit, flux_unit, trans_unit):
     if test_spec[2] is not None:
         mask *= trans_unit
 
-    rv = Q.rv_precision(wav, flux, mask)
+    rv = rv_precision(wav, flux, mask)
     assert rv.unit == m_per_s
     assert not hasattr(rv.value, "__len__")  # assert value is a scalar
     assert isinstance(rv, u.Quantity)
@@ -43,7 +50,7 @@ def test_rvprev_calc_with_lists(test_spec):
     wav = list(test_spec[0])
     flux = list(test_spec[1])
     mask = test_spec[2]
-    rv = Q.rv_precision(wav, flux, mask)
+    rv = rv_precision(wav, flux, mask)
     assert not hasattr(rv.value, "__len__")  # assert value is a scalar
     assert isinstance(rv, u.Quantity)
     assert rv.unit == m_per_s
@@ -53,7 +60,7 @@ def test_sqrt_sum_wis_with_no_units(test_spec):
     """Test that sqrt_sum_wis can handle inputs as Quantities or unitless.
     Returns a dimensionless unscaled Quantity.
     """
-    sqrtsumwis = Q.sqrt_sum_wis(test_spec[0], test_spec[1], test_spec[2])
+    sqrtsumwis = sqrt_sum_wis(test_spec[0], test_spec[1], test_spec[2])
     # Doesn't turn into quantity if does not have to.
     assert not isinstance(sqrtsumwis, u.Quantity)
     assert not hasattr(sqrtsumwis, "__len__")  # assert value is a scalar
@@ -70,7 +77,7 @@ def test_sqrt_sum_wis(test_spec, wav_unit, flux_unit, trans_unit):
     if test_spec[2] is not None:
         mask *= trans_unit
 
-    sqrtsumwis = Q.sqrt_sum_wis(wav, flux, mask)
+    sqrtsumwis = sqrt_sum_wis(wav, flux, mask)
     print("wav", wav, type(wav))
     print("flux", flux, type(flux))
     print("mask", mask, type(mask))
@@ -97,7 +104,7 @@ def test_relation_of_rv_to_sqrtsumwis(test_spec, wav_unit, flux_unit, trans_unit
         mask *= trans_unit
         mask = mask ** 2
     assert np.all(
-        Q.rv_precision(wav, flux, mask=mask) == c / Q.sqrt_sum_wis(wav, flux, mask=mask)
+        rv_precision(wav, flux, mask=mask) == c / sqrt_sum_wis(wav, flux, mask=mask)
     )
 
 
@@ -109,11 +116,11 @@ def test_transmission_reduces_precision(test_spec):
 
     # Value should be less then normal if trans <=1
     if transmission is not None:
-        assert Q.rv_precision(wav, flux, mask=None) < Q.rv_precision(
+        assert rv_precision(wav, flux, mask=None) < rv_precision(
             wav, flux, mask=transmission
         )
[…]
