### 2019-03-08 aeb5cde
GridError -> ValueError
Former-commit-id: f5f783e031455f1a0725378a7e9aaae13598ac33 [formerly 3415ac16947bc461bbfec791b0625e89866df51a]
Former-commit-id: 1b44114527c967c5d1e97f3092c07e703f8b0386
