+++ tests/test_utilities.py
@@ -437,16 +437,6 @@ def test_load_btsettl_spectrum(photons):
 
 
 @xfail(raises=ModuleNotFoundError, reason="Issue with Starfish install.")
-@pytest.mark.parametrize("params", [[8000, 4.5, 0, 0], [3900, 0.5, 0, 0]])
-def test_invalid_load_btsettl_spectrum(params):
-    # Invalid CIFIST parameters
-    from Starfish.constants import GridError
-
-    with pytest.raises(GridError):
-        load_btsettl_spectrum(params, wl_range=[21000, 22000])
-
-
-@xfail(raises=ModuleNotFoundError, reason="Issue with Starfish install.")
 @pytest.mark.parametrize("params", [[3900, 4.5, 1, 0], [3900, 4.5, 0, 1]])
 def test_invalid_feh_alpha_load_btsettl_spectrum(params):
     # Invalid CIFIST parameters
@@ -475,8 +465,12 @@ def test_load_aces_spectrum(photons):
     ],
 )
 def test_invalid_load_aces_spectrum(params):
-    # Invalid CIFIST parameters
-    from Starfish.constants import GridError
-
-    with pytest.raises(GridError):
+    with pytest.raises(ValueError):
         load_aces_spectrum(params, wl_range=[21000, 22000])
+
+
+@xfail(raises=ModuleNotFoundError, reason="Issue with Starfish install.")
+@pytest.mark.parametrize("params", [[8000, 4.5, 0, 0], [3900, 0.5, 0, 0]])
+def test_invalid_load_btsettl_spectrum(params):
+    with pytest.raises(ValueError):
+        load_btsettl_spectrum(params, wl_range=[21000, 22000])
