### 2019-03-08 04b9acc
Remove starfish install xfail.
Former-commit-id: fd600ed94a868b9fb149fe826e5a606f75a57539 [formerly ad61cbfbb8cf5119415ca5d9c7daf65599c06e8a]
Former-commit-id: 9a98cde02a9e4841fbd5376d822b5b141ba744af
