### 2019-03-08 2f19317
info rstcheck error level
Former-commit-id: 5e7b36ba28e31341c9fcda45923c9392755ae7c9 [formerly a24d70132035eb7baab684f3f5da86973eb1149c]
Former-commit-id: b1b0b2cf4cb24a51274747cfbbf52ba34b5596df
