### 2019-03-08 8ca7eab
Revert "Remove starfish install xfail."
This reverts commit 1a283e1249f5863d2c0e57447562b083d36d43c8.
