### 2019-03-08 8a4e8a4
Revert "Remove starfish install xfail."
This reverts commit 04b9acce507ad5f70e4f2ded57fafdbfb129a4d7 [formerly 9a98cde02a9e4841fbd5376d822b5b141ba744af] [formerly fd600ed94a868b9fb149fe826e5a606f75a57539 [formerly ad61cbfbb8cf5119415ca5d9c7daf65599c06e8a]].


Former-commit-id: a812c51fd8d800a840040808a7c49bf837d0bfef [formerly d68c4d76512898052fecee878bc2d5bf57f8ee15]
Former-commit-id: 091e904b091f3fdd3044f530d321b71defe25bab
