+++ tests/test_utilities.py
@@ -427,6 +427,7 @@ def test_doppler_limits_rv_0(wmin, wmax):
     assert new_max == wmax
 
 
+@xfail(raises=ModuleNotFoundError, reason="Issue with Starfish install.")
 @pytest.mark.parametrize("photons", [True, False])
 def test_load_btsettl_spectrum(photons):
     wav, flux = load_btsettl_spectrum(
@@ -435,6 +436,7 @@ def test_load_btsettl_spectrum(photons):
     assert len(wav) == len(flux)
 
 
+@xfail(raises=ModuleNotFoundError, reason="Issue with Starfish install.")
 @pytest.mark.parametrize("params", [[3900, 4.5, 1, 0], [3900, 4.5, 0, 1]])
 def test_invalid_feh_alpha_load_btsettl_spectrum(params):
     # Invalid CIFIST parameters
@@ -442,6 +444,7 @@ def test_invalid_feh_alpha_load_btsettl_spectrum(params):
         load_btsettl_spectrum(params, wl_range=[21000, 22000])
 
 
+@xfail(raises=ModuleNotFoundError, reason="Issue with Starfish install.")
 @pytest.mark.parametrize("photons", [True, False])
 def test_load_aces_spectrum(photons):
     wav, flux = load_aces_spectrum(
@@ -450,6 +453,7 @@ def test_load_aces_spectrum(photons):
     assert len(wav) == len(flux)
 
 
+@xfail(raises=ModuleNotFoundError, reason="Issue with Starfish install.")
 @pytest.mark.parametrize(
     "params",
     [
@@ -465,6 +469,7 @@ def test_invalid_load_aces_spectrum(params):
         load_aces_spectrum(params, wl_range=[21000, 22000])
 
 
+@xfail(raises=ModuleNotFoundError, reason="Issue with Starfish install.")
 @pytest.mark.parametrize("params", [[8000, 4.5, 0, 0], [3900, 0.5, 0, 0]])
 def test_invalid_load_btsettl_spectrum(params):
     with pytest.raises(ValueError):
