+++ docs/source/theoretical_precision.rst
@@ -1,4 +1,4 @@
-
+******************************************
 Theoretical Precision of Synthetic Spectra
 ******************************************
 
