### 2019-03-08 48a5535
Refactor rst file names.
