### 2019-03-09 dc1d621
Remove unnessesary config files, and rename one to tests/test_config.yaml
