+++ tests/test_config.yaml
@@ -26,9 +26,7 @@ bands:
 
 # Add your custom band limits here
 custom_bands:
-    TEST: [2.1,2.2]
-
-
+  TEST: [2.1, 2.2]
 
 ## Starfish Keywords
 # These keywords are necessary to exist in the config.yaml (to enable Starfish to run)
@@ -42,17 +40,17 @@ custom_bands:
 # above.
 # If you are wanting to use "Eniric" with the other features "Starfish" good luck!
 # (or open an issue github issue).
-name: ""
+name: "test config"
 outdir: ""
 plotdir: ""
 grid:
-        raw_path: ""
-        hdf5_path: ""
-        key_name: ""
-        parname: []
-        parrange: []
-        wl_range: []
-        buffer: 0
+  raw_path: ""
+  hdf5_path: ""
+  key_name: ""
+  parname: []
+  parrange: []
+  wl_range: []
+  buffer: 0
 data:
-    grid_name: ""
-    instruments: []
+  grid_name: ""
+  instruments: []
