### 2019-03-09 3424adb
Don't move test config file in ci service configurations
