### 2019-03-09 94ab6d5
custom bands must be lists, fix path spelling
