### 2019-03-09 4d1d951
Updated config that allows for in memory changes and rewriting config file.
Updated tests.
yaml loaded/stored as orderedDict to keep yaml order.
