+++ eniric/_config.py
@@ -1,48 +1,129 @@
-import yaml
+# Inspired by Starfish
+
 import os
+import shutil
+
+import oyaml as yaml
+
+from . import DEFAULT_CONFIG_FILE
 
 
-class Config:
-    """
-    This is a class to allow lazy-loading of config files. This means that the atributes are only read from the file
-    when accessed in the code, allowing for changes in the file without having to restart the python instance
-    """
+class Config(object):
+    def __init__(self, path):
+        """Creates a persistent Config object from the give file.
 
-    def __init__(self, filename):
-        self.filename = filename
-        # Format string for saving/reading orders
-        self.specfmt = "s{}_o{}"
+        This class is not meant to be instantiated by the User,
+        but rather to be used via ``eniric.config``.
 
-    def __getitem__(self, item):
-        with open(self.filename) as f:
-            base = yaml.safe_load(f)
-            return base[item]
+        Parameters
+        ----------
+        path: str or path-like
+            The filename for creating the Config. Must be YAML.
+        """
+        self._path = path
+        self._protect_rewrites = True
 
-    def __getattribute__(self, item):
-        # Short circuit on filename or else infinite recursion
-        if item == "filename":
-            return super().__getattribute__(item)
-        with open(self.filename) as f:
-            base = yaml.safe_load(f)
-            if item == "paths":
-                paths = base["paths"]
-                for key, value in paths.items():
+        with open(self._path, "r") as fd:
+            self._config = yaml.safe_load(fd)
+
+        self._protect_rewrites = os.path.abspath(path) == DEFAULT_CONFIG_FILE
+
+    def __contains__(self, key):
+        return key in self._config
+
+    def __delitem__(self, key):
+        del self._config[key]
+
+    def __eq__(self, other):
+        return self._config.__eq__(other)
+
+    def __getitem__(self, key):
+        return self._config[key]
+
+    def __setitem__(self, key, value):
+        ret = self._config.__setitem__(key, value)
+        return ret
+
+    def __getattr__(self, key):
+        if key in self:
+            if key == "paths":
+                paths = self["paths"]
+                for k, value in paths.items():
                     if isinstance(value, list):
-                        paths[key] = os.path.join(*value)
+                        paths[k] = os.path.join(*value)
                 return paths
-            elif item == "bands":
-                return base["bands"]
-            elif item == "custom_bands":
-                return base["custom_bands"]
-            elif item == "atmmodel":
-                return base["atmmodel"]
-            elif item == "cache":
-                cache = base["cache"]
+            elif key == "cache":
+                cache = self["cache"]
                 if (cache["location"] is None) or (cache["location"] == "None"):
                     cache["location"] = None  # Disables caching
-                else:
-                    if isinstance(cache["location"], list):
-                        cache["location"] = os.path.join(*cache["location"])
+                elif isinstance(cache["location"], list):
+                    cache["location"] = os.path.join(*cache["location"])
                 return cache
-            else:
-                return super().__getattribute__(item)
+            return self[key]
+        else:
+            return super().__getattribute__(key)
+
+    def __setattr__(self, key, value):
+        if key not in ["_path", "_protect_rewrites", "_config"]:
+            if key in self:
+                self.__setitem__(key, value)
+                self._rewrite()
+
+        super().__setattr__(key, value)
+
+    def _rewrite(self):
+        if self._protect_rewrites:
+            raise RuntimeError(
+                "The default file is not allowed to be overwritten. Please copy a file using "
+                "config.copy_file(<path>) for your use."
+            )
+        with open(self._path, "w") as fd:
+            yaml.safe_dump(self._config, fd)
+
+    def update(self, d={}, **kwargs):
[…]
