+++ .isort.cfg
@@ -1,5 +1,5 @@
 [settings]
-known_third_party = Starfish,astropy,hypothesis,joblib,matplotlib,multiprocess,numpy,pandas,pytest,setuptools,tqdm,yaml
+known_third_party = Starfish,astropy,hypothesis,joblib,matplotlib,multiprocess,numpy,oyaml,pandas,pytest,setuptools,tqdm,yaml
 multi_line_output=3
 include_trailing_comma=True
 force_grid_wrap=0
