### 2019-03-10 006e016
Use 'from eniric import config'
