+++ eniric/obsolete/nIRanalysis.py
@@ -10,15 +10,16 @@ from numpy import ndarray
 
 import eniric
 import eniric.io_module as io
+from eniric import config
 from eniric.broaden import convolution
 from eniric.obsolete.utilities import read_spectrum
 
 # set stuff for latex usage
 rc("text", usetex=True)
 
-data_rep = eniric.config.paths["phoenix_dat"]
-results_dir = eniric.config.paths["results"]
-resampled_dir = eniric.config.paths["resampled"]
+data_rep = config.paths["phoenix_dat"]
+results_dir = config.paths["results"]
+resampled_dir = config.paths["resampled"]
 
 # models form PHOENIX-ACES
 M0_ACES = data_rep + "lte03900-4.50-0.0.PHOENIX-ACES-AGSS-COND-2011-HiRes_wave.dat"

[…]
