### 2019-03-10 5bf6d32
Move test data to tests/data
