+++ tests/test_config.yaml
@@ -1,7 +1,7 @@
 paths:
-  phoenix_raw: [data, test_data, phoenix-raw]
-  btsettl_raw: [data, test_data, btsettl-raw]
-  test_data: [data, test_data]
+  phoenix_raw: [tests, data, phoenix-raw]
+  btsettl_raw: [tests, data, btsettl-raw]
+  test_data: [tests, data]
   atmmodel: [data, atmmodel]
   precision: [precision]
   phoenix_dat: [data, test_data, obsolete, PHOENIX-ACES_spectra]
