### 2019-03-11 6833246
Add pathdir property to config. Make config paths relative to config file location. Refactor os.path.join -> join
