+++ eniric/config.yaml
@@ -6,7 +6,7 @@
 paths:
   phoenix_raw: ["..", "data", "phoenix-raw"] # Path to Phoenix-ACES spectra
   btsettl_raw: ["..", "data", "btsettl-raw"] # Path to BT-Settl spectra
-  test_data: ["..", "data", "test_data"]
+  test_data: ["..", "tests", "data"]
   atmmodel: ["..", "data", "atmmodel"]
   precision: ["precision"]  # A place to put precision results
   # Obsolete paths for testing only
