+++ eniric/config.yaml
@@ -8,7 +8,7 @@ paths:
   btsettl_raw: ["..", "data", "btsettl-raw"] # Path to BT-Settl spectra
   test_data: ["..", "tests", "data"]
   atmmodel: ["..", "data", "atmmodel"]
-  precision: ["precision"]  # A place to put precision results
+  precision: ["..", "data", "precision"]  # A place to put precision results
   # Obsolete paths for testing only
   phoenix_dat: ["data", "test_data", "obsolete", "PHOENIX-ACES_spectra"]
   results:  ["..", "data", "test_data", "obsolete", "results"]
