### 2019-03-11 e9820ba
Remove resampled_data dependecy, use test_spectrum instead
