### 2019-03-11 4e10adc
Remove obsolete data, scripts, functions (#136)
* Remove obsolete modules, function, config and script calls.

* Remove resampled_data dependency, use test_spectrum instead

* xfail Starfish error
