+++ eniric/config.yaml
@@ -8,12 +8,7 @@ paths:
   btsettl_raw: ["..", "data", "btsettl-raw"] # Path to BT-Settl spectra
   test_data: ["..", "data", "test_data"]
   atmmodel: ["..", "data", "atmmodel"]
-
   precision: ["precision"]  # A place to put precision results
-  # Obsolete paths for testing only
-  phoenix_dat: ["data", "test_data", "obsolete", "PHOENIX-ACES_spectra"]
-  results:  ["..", "data", "test_data", "obsolete", "results"]
-  resampled: ["..", "data", "test_data", "obsolete", "resampled"]
 
 # Wavelength bands.
 bands:
