+++ .travis.yml
@@ -24,8 +24,6 @@ before_script:
 
   # Download the Data for eniric and testing
   - download_eniric_data.sh  # Some files from Dropbox
-  # Generate data
-  - make_test_data.py
 
   # Prepare atmosphere models
   - make atmos V=1

[…]
