### 2019-03-11 4535ec9
Remove obsolete data, scripts, functions (#136)
* Remove obsolete modules, function, config and script calls.

* Remove resampled_data dependency, use test_spectrum instead

* xfail Starfish error


Former-commit-id: 42b6057db1e48c1f0768874cab8bf726b391818e [formerly bcc2be2be6dae3fcd93ae09df1439128a9749663]
Former-commit-id: 4ad061a920cdd95538c8cf2d795dd1a4ac6dace4
