### 2019-03-12 416ec1b
Fixture to change config file to test_config.yaml
