### 2019-03-12 a2b781c
Editable ci installs.
