+++ .travis.yml
@@ -16,7 +16,7 @@ install:
   - travis_retry pip install -r requirements.txt
 
   # Now install eniric
-  - pip install .[dev,test,docs,ci]
+  - pip install -e .[dev,test,docs,ci]
 
 before_script:
   - export MPLBACKEND="AGG"  # Non-interactive
