+++ .shippable.yml
@@ -16,7 +16,7 @@ build:
     - travis_retry pip install -r requirements.txt
 
     # Now install eniric
-    - pip install .[dev,test,docs,ci]
+    - pip install -e .[dev,test,docs,ci]
 
     # Setup config and make test data
     - export MPLBACKEND="AGG"  # Non-interactive
