+++ appveyor.yml
@@ -16,7 +16,7 @@ install:
   - "%PYTHON%\\python.exe -m pip install -r requirements.txt"
 
   # Installing eniric now
-  - "%PYTHON%\\python.exe -m pip install .[dev,test,docs,ci]"
+  - "%PYTHON%\\python.exe -m pip install -e .[dev,test,docs,ci]"
 
   # Setup config and make data
   # This is not the best place to put this but before_test didn't work.
