+++ tests/config.yaml
@@ -1,54 +1,0 @@
-# YAML configuration script for Eniric
-
-# Paths to locations used by eniric
-paths:
-  phoenix_raw: ["data", "test_data", "phoenix-raw"]
-  btsettl_raw: ["data", "test_data", "btsettl-raw"]
-  test_data: ["data","test_data"]
-  atmmodel: ["data", "atmmodel"]
-  precision: ["precision"]
-
-cache:
-  # location: None  # Disables cache
-  location: [".joblib"]
-
-atmmodel:
-  # Base name of telluric model
-  base: "Average_TAPAS_2014"
-
-# Wavelength bands.
-bands:
-  all: ["K", "H", "J", "Y", "Z", "TEST"]
-
-# Add your custom band limits here
-custom_bands:
-  TEST: [2.1, 2.2]
-
-
-
-## Starfish Keywords
-# These keywords are necessary to exist in the config.yaml (to enable Starfish to run)
-# but are not used by and have no affect on Eniric and the loading of the phoenix spectral libraries.
-# As such they are all set to nothing.
-#
-# To configure the path to the phoenix libraries use
-#    paths:
-#        phoenix_raw:
-#        btsettl_raw:
-# above.
-# If you are wanting to use "Eniric" with the other features "Starfish" good luck!
-# (or open an issue github issue).
-name: ""
-outdir: ""
-plotdir: ""
-grid:
-  raw_path: ""
-  hdf5_path: ""
-  key_name: ""
-  parname: []
-  parrange: []
-  wl_range: []
-  buffer: 0
-data:
-  grid_name: ""
-  instruments: []
