### 2019-03-12 e254ee1
Update config. (#137)
* Remove unnessesary config files, and rename one to tests/test_config.yaml

* Updated config that allows for in memory changes and rewriting config file.

* Don't move test config file in ci service configurations

* custom bands must be lists,

* Set DEFAULT_CONFIG_FILE in __init__.

* Move and use dumped test_config

Updated tests.
yaml loaded/stored as orderedDict to keep yaml order.

* Use 'from eniric import config'

* Move test data to tests/data

* Add pathdir property to config.
Make config paths relative to config file location.
Refactor os.path.join -> join

* 
[…]
