+++ tests/test_config.py
@@ -3,60 +3,138 @@ import os
 import pytest
 import yaml
 
-from eniric import config
+from eniric import DEFAULT_CONFIG_FILE, config
+from eniric._config import Config
 
-default_config = os.path.join(os.path.dirname(os.path.dirname(__file__)), "config.yaml")
+base_dir = os.path.dirname(__file__)
+test_filename = os.path.join(base_dir, "data", "test_config.yaml")
 
 
 class TestConfig:
-    @pytest.mark.xfail()
+    @pytest.fixture
+    def test_config(self):
+        """Config file for testing."""
+        filename = test_filename
+        yield Config(filename)
+
     def test_default_filename(self):
-        assert config.filename == default_config
+        default_config = os.path.join(
+            os.path.dirname(os.path.dirname(__file__)), "eniric", "config.yaml"
+        )
+        assert DEFAULT_CONFIG_FILE == default_config
+
+    def test_base_dots(self):
+        assert config.paths == config["paths"]
+        assert config.cache == config["cache"]
+        assert config.atmmodel == config["atmmodel"]
+        assert config.bands == config["bands"]
+        assert config.custom_bands == config["custom_bands"]
 
     @pytest.mark.parametrize(
-        "key, value",
+        "key, values",
         [
-            ("phoenix_raw", os.path.join(*["data", "test_data", "phoenix-raw"])),
-            ("btsettl_raw", os.path.join(*["data", "test_data", "btsettl-raw"])),
-            ("atmmodel", os.path.join(*["data", "atmmodel"])),
-            ("precision", "precision"),
-            ("test_data", os.path.join(*["data", "test_data"])),
+            ("phoenix_raw", ["..", "data", "phoenix-raw"]),
+            ("btsettl_raw", ["..", "data", "btsettl-raw"]),
+            ("atmmodel", ["..", "data", "atmmodel"]),
+            ("precision", ["..", "data", "precision"]),
         ],
     )
-    def test_paths_keys(self, key, value):
-        assert config.paths[key] == value
+    def test_default_paths_keys(self, key, values):
+        assert config.paths[key] == os.path.join(*values)
+
+    @pytest.mark.parametrize(
+        "key, values",
+        [
+            ("phoenix_raw", ["phoenix-raw"]),
+            ("btsettl_raw", ["btsettl-raw"]),
+            ("atmmodel", ["..", "..", "data", "atmmodel"]),
+            ("precision", ["..", "..", "data", "precision"]),
+        ],
+    )
+    def test_paths_keys(self, test_config, key, values):
+        assert test_config.paths[key] == os.path.join(*values)
 
     def test_paths(self):
         assert isinstance(config.paths, dict)
 
-    def test_cache(self):
+    def test_cache(self,):
         assert isinstance(config.cache, dict)
         assert config.cache["location"] == ".joblib"
 
-    def test_atmmodel(self):
+    def test_atmmodel(self, test_config):
         assert isinstance(config.atmmodel, dict)
         assert config.atmmodel["base"] == "Average_TAPAS_2014"
 
-    def test_bands(self):
+    def test_default_bands(self):
         assert isinstance(config.bands, dict)
         assert isinstance(config.bands["all"], list)
-        assert "K" in config.bands["all"]
-        assert "J" in config.bands["all"]
+        assert config.bands["all"] == [
+            "VIS",
+            "GAP",
+            "Z",
+            "Y",
+            "J",
+            "H",
+            "K",
+            "CONT",
+            "NIR",
+            "TEST",
+        ]
+
+    def test_bands(self, test_config):
+        assert test_config.bands["all"] == ["K", "H", "J", "Y", "Z", "TEST"]
 
     def test_custom_bands(self):
         assert isinstance(config.custom_bands, dict)
         for value in config.custom_bands.values():
[…]
