+++ tests/test_utilities.py
@@ -429,21 +429,11 @@ def test_doppler_limits_rv_0(wmin, wmax):
 
 @xfail(raises=ModuleNotFoundError, reason="Issue with Starfish install.")
 @pytest.mark.parametrize("photons", [True, False])
-def test_load_btsettl_spectrum(photons):
+def test_load_btsettl_spectrum(photons, use_test_config):
     wav, flux = load_btsettl_spectrum(
         [3900, 4.5, 0, 0], photons=photons, air=False, wl_range=[21000, 22000]
     )
     assert len(wav) == len(flux)
-
-
-@xfail(raises=ModuleNotFoundError, reason="Issue with Starfish install.")
-@pytest.mark.parametrize("params", [[8000, 4.5, 0, 0], [3900, 0.5, 0, 0]])
-def test_invalid_load_btsettl_spectrum(params):
-    # Invalid CIFIST parameters
-    from Starfish.constants import GridError
-
-    with pytest.raises(GridError):
-        load_btsettl_spectrum(params, wl_range=[21000, 22000])
 
 
 @xfail(raises=ModuleNotFoundError, reason="Issue with Starfish install.")
@@ -456,7 +446,7 @@ def test_invalid_feh_alpha_load_btsettl_spectrum(params):
 
 @xfail(raises=ModuleNotFoundError, reason="Issue with Starfish install.")
 @pytest.mark.parametrize("photons", [True, False])
-def test_load_aces_spectrum(photons):
+def test_load_aces_spectrum(photons, use_test_config):
     wav, flux = load_aces_spectrum(
         [3900, 4.5, 0, 0], photons=photons, air=False, wl_range=[21000, 22000]
     )
@@ -475,8 +465,12 @@ def test_load_aces_spectrum(photons):
     ],
 )
 def test_invalid_load_aces_spectrum(params):
-    # Invalid CIFIST parameters
-    from Starfish.constants import GridError
-
-    with pytest.raises(GridError):
+    with pytest.raises(ValueError):
         load_aces_spectrum(params, wl_range=[21000, 22000])
+
+
+@xfail(raises=ModuleNotFoundError, reason="Issue with Starfish install.")
+@pytest.mark.parametrize("params", [[8000, 4.5, 0, 0], [3900, 0.5, 0, 0]])
+def test_invalid_load_btsettl_spectrum(params):
+    with pytest.raises(ValueError):
+        load_btsettl_spectrum(params, wl_range=[21000, 22000])
