+++ .shippable.yml
@@ -16,11 +16,11 @@ build:
     - travis_retry pip install -r requirements.txt
 
     # Now install eniric
-    - pip install .[dev,test,docs,ci]
+    - pip install -e .[dev,test,docs,ci]
 
     # Setup config and make test data
     - export MPLBACKEND="AGG"  # Non-interactive
-    - mv tests/config.yaml config.yaml
+    # - mv tests/config.yaml config.yaml
 
     # Download the Data for eniric and testing
     - download_eniric_data.sh  # Some files from Dropbox

[…]
