### 2019-03-13 971b2be
Add Python/Scripts to appveryor config.
