+++ appveyor.yml
@@ -12,6 +12,7 @@ cache:
   - "data\\test_data -> appveyor.yml"           # Use cache unless appveyor.yml changes.
 
 install:
+  - SET PATH=%PYTHON%;%PYTHON%\Scripts\;%path%
   - "%PYTHON%\\python.exe -m pip install --upgrade pip"
   # We need wheel installed to build wheels
   - "%PYTHON%\\python.exe -m pip install wheel"
