+++ docs/source/examples.rst
@@ -9,25 +9,25 @@ Below are the links to their location in github or rendered with `nbveiwer <http
 These are:
 
 
-* Atmosphere model (`Github <https://github.com/jason-neal/eniric/blob/master/docs/Notebooks/atmosphere_example.ipynb>`_) (`nbviewer <https://nbviewer.jupyter.org/github/jason-neal/eniric/blob/master/docs/Notebooks/Compare_precisions.ipynb>`_)
-* Compare precisions (`Github <https://github.com/jason-neal/eniric/blob/master/docs/Notebooks/Compare_precisions.ipynb>`_) (`nbviewer <https://nbviewer.jupyter.org/github/jason-neal/eniric/blob/master/docs/Notebooks/Compare_precisions.ipynb>`_)
+* Atmosphere model (`Github <https://github.com/jason-neal/eniric/blob/master/docs/Notebooks/atmosphere_example.ipynb>`__) (`nbviewer <https://nbviewer.jupyter.org/github/jason-neal/eniric/blob/master/docs/Notebooks/Compare_precisions.ipynb>`__)
+* Compare precisions (`Github <https://github.com/jason-neal/eniric/blob/master/docs/Notebooks/Compare_precisions.ipynb>`__) (`nbviewer <https://nbviewer.jupyter.org/github/jason-neal/eniric/blob/master/docs/Notebooks/Compare_precisions.ipynb>`__)
 
-* Incremental_precision.ipynb (`Github <https://github.com/jason-neal/eniric/blob/master/docs/Notebooks/Incremental_precision.ipynb>`_) (`nbviewer <https://nbviewer.jupyter.org/github/jason-neal/eniric/blob/master/docs/Notebooks/Incremental_precision.ipynb>`_)
-* Precision_example.ipynb (`Github <https://github.com/jason-neal/eniric/blob/master/docs/Notebooks/Precision_example.ipynb>`_) (`nbviewer <https://nbviewer.jupyter.org/github/jason-neal/eniric/blob/master/docs/Notebooks/Precision_example.ipynb>`_)
-* Precison_with_doppler-K-band.ipynb (`Github <https://github.com/jason-neal/eniric/blob/master/docs/Notebooks/Precison_with_doppler-K-band.ipynb>`_) (`nbviewer <https://nbviewer.jupyter.org/github/jason-neal/eniric/blob/master/docs/Notebooks/Precison_with_doppler-K-band.ipynb>`_)
-* Precison_with_doppler-Z-band.ipynb (`Github <https://github.com/jason-neal/eniric/blob/master/docs/Notebooks/Precison_with_doppler-Z-band.ipynb>`_) (`nbviewer <https://nbviewer.jupyter.org/github/jason-neal/eniric/blob/master/docs/Notebooks/Precison_with_doppler-Z-band.ipynb>`_)
-* Solar_precision.ipynb (`Github <https://github.com/jason-neal/eniric/blob/master/docs/Notebooks/Solar_precision.ipynb>`_) (`nbviewer <https://nbviewer.jupyter.org/github/jason-neal/eniric/blob/master/docs/Notebooks/Solar_precision.ipynb>`_)
-* Effect of spectral gradient (`Github <https://github.com/jason-neal/eniric/blob/master/docs/Notebooks/spectral_gradient.ipynb>`_) (`nbviewer <https://nbviewer.jupyter.org/github/jason-neal/eniric/blob/master/docs/Notebooks/spectral_gradient.ipynb>`_)
[…]
