### 2019-03-13 aaad645
Add rstcheck.cfg to specify settings
Former-commit-id: 2387ad74709d692c6f0ed694034918b487d2c5c7 [formerly f0001f5d89582ab389062bd6391ef7e34fd5d5fc]
Former-commit-id: d89b35ad8db92fad495b4e82a082ba19d969d6c1
