+++ docs/.rstcheck.cfg
@@ -1,0 +1,5 @@
+[rstcheck]
+ignore_directives=
+ignore_roles=
+ignore_messages=(Unknown directive type "automodule".|Unknown directive type "autofunction".|No directive entry for "autofunction" in module "docutils.parsers.rst.languages.en".|No directive entry for "automodule" in module "docutils.parsers.rst.languages.en".)
+report=warning
