### 2019-03-13 bed8fdc
Reduce appveyor test environments
