### 2019-03-13 31eb552
adjust scripts/__init__.py
