+++ scripts/__init__.py
@@ -1,6 +1,8 @@
 __all__ = [
-    "barycenter_broaden_atmmodel.py",
+    "barycenter_broaden_atmmodel",
+    "csv2tsv",
     "phoenix_precision",
     "precision_four_panel",
     "split_atmmodel",
+    "untar_here",
 ]
