### 2019-03-13 a70c295
Remove bash specification on code block
Former-commit-id: a8d2701f307bc061bad8178cb6148cecd1b40013 [formerly 7e1792a9a9d23abc1a742da9d047c7992b39bcbd]
Former-commit-id: 8b24e599f7ade3b02d505b3f498f4d02ba6bcd1c
