+++ eniric/broaden.py
@@ -128,7 +128,7 @@ def rotational_convolution(
                 convolved_flux = np.asarray(
                     num_procs(delayed(element_rot_convolution)(wav) for wav in tqdm_wav)
                 )
-            except AttributeError:
+            except TypeError:
                 raise TypeError(
                     "num_proc must be an int, joblib.parallel.Parallel. Not '{}'".format(
                         type(num_procs)
@@ -239,7 +239,7 @@ def resolution_convolution(
             convolved_flux = np.asarray(
                 num_procs(delayed(element_res_convolution)(wav) for wav in tqdm_wav)
             )
-        except AttributeError:
+        except TypeError:
             raise TypeError(
                 "num_proc must be an int or joblib.parallel.Parallel. Not '{}'".format(
                     type(num_procs)
