### 2019-03-14 543b0cb
Use Joblib parallel rather than multiprocessing. (#138)
* Replace multiprocess with joblib as parallelization creation inside convolution.

* Open/close file for every line written.

* Joblib parallel valid input for convlutions.

* Use asarray

* Simplify normalization / return

* Fix TypeError and  match message.
