### 2019-03-14 b3ff564
Update phoenix-aces data download. (#131)
* Use http for phoenix data

* Move download scripts to scripts/download.

Create new download script using new Starfish functions.

Cache data on CI services

* Fix starfish #egg

* Fix download_test_aces.py Starfish downloaded takes a iterable of iterables.

* Fix appveyor escape characters in paths \\.

* fix ps_download_eniric_data.ps1

 Add Python/Scripts path to appveryor config.


Former-commit-id: 5d57495b073cfd4296c6aeff9005ba06f3244556 [formerly c342965d6de0c6a646199e615452f6a886817878]
Former-commit-id: 81c66a359e70623314f13ae6b78ab8ad5b5a6d
[…]
