### 2019-03-14 21902eb
Update phoenix-aces data download. (#131)
* Use http for phoenix data

* Move download scripts to scripts/download.

Create new download script using new Starfish functions.

Cache data on CI services

* Fix starfish #egg

* Fix download_test_aces.py Starfish downloaded takes a iterable of iterables.

* Fix appveyor escape characters in paths \\.

* fix ps_download_eniric_data.ps1

 Add Python/Scripts path to appveryor config.
