+++ .travis.yml
@@ -8,6 +8,7 @@ python:
 cache:
   directories:
     - $HOME/.cache/pip
+    - $HOME/data/test_data
 
 install:
   - pip install --upgrade pip
