### 2019-03-14 2f5c3a8
Remove/change some slow tests
Former-commit-id: b2649b796caa3d92c20be8a9141c3e1bee130f0d [formerly 094a016e6489eb1e1bbbc44f9a2d2c3156dbded9]
Former-commit-id: 58fceae9bacac1be786f16c4ba17cad9904c8ded
