### 2019-03-17 91aa2bb
Include data downloading.
Related to #117


Former-commit-id: 28954df0ba3eaf90ad7cffa224032a3315676a47 [formerly 4cf5679b0ed44811703b37f202a602d898f4bb7d]
Former-commit-id: 3978d7329849dc97788bdd90fe76a89aa68cf63a
