### 2019-03-17 b720448
Include data downloading.
Related to #117
