### 2019-03-28 jason-neal/eniric#114 @jason-neal
Have move documentation all to read the docs.
Updated the installation and configuration documentation.
