### 2019-03-29 #1053 @jason-neal
I have managed to finish documentation on read the docs https://eniric.readthedocs.io/en/latest/

The latest version is currently still in the develop branch but will merge into master and change to v1.0 once it is approved.
