### 2019-04-10 #1053 @arfon
@jason-neal - could you merge this PR? It will fix the citation formatting for multiple references: https://github.com/jason-neal/eniric/pull/141
