### 2019-04-10 #1053 @arfon
@jason-neal - this is looking pretty good to me. The only issue I see is with the figure placement. Do you think you could try moving the figure up nearer the top of the paper, that way it might not end up floating in the references section at the end of the paper.
