### 2019-04-23 #1053 @crawfordsm
@jason-neal You have done a really great job of updating the documentation along with the notebooks!  It really makes the package much more accessible and now I really feel I can jump in to using it with the clear API.

I recommend Eniric for publication in JOSS.
