### 2019-04-23 #1053 @arfon
> I recommend Eniric for publication in JOSS.

👍 thanks @crawfordsm.
