### 2019-05-03 #1053 @arfon
Thanks @jason-neal - one more thing - could you please update the authors on the Zenodo archive to match those of the paper? I'm currently listed as an author 😁
