### 2019-05-06 #1053 @jason-neal
@afron Do you know how to do that? I have edited the authors on Zenodo and saved it but it does not seem to affect the shown authors. Do I need to **publish** it to make the change?
