### 2019-05-06 #1053 @arfon
> @afron Do you know how to do that? I have edited the authors on Zenodo and saved it but it does not seem to affect the shown authors. Do I need to publish it to make the change?

I'm not sure sorry. If there's a **publish** button then perhaps click it?
