### 2019-05-07 #1053 @arfon
```
- 10/gfdcsn is INVALID
- 10/gfdcsp is INVALID
- 10/gfdcsg is INVALID
- 10/gfdcsf is INVALID
- 10/cqg52h is INVALID
- 10/gfdcst is INVALID
- 10/gfdcsz is INVALID
- 10/dk66bp is INVALID
- 10/gdmwfg is INVALID
- 10/gddxj4 is INVALID
- 12.1002/asna.201312004 is INVALID
```

@jason-neal please fix up these DOIs ☝️in your bibtex file.
